imxrt_ral::bee
Module CTRL

BEE Control Register

Modules

AC_PROT_EN

Enable access permission control When AC_PROT_EN is asserted, all encrypted regions are limited to be ARM core access only

AC_PROT_EN_LOCK

Lock bit for ac_prot

BEE_ENABLE

BEE enable bit

BEE_ENABLE_LOCK

Lock bit for bee_enable

CTRL_AES_MODE_R0

AES mode of region0

CTRL_AES_MODE_R0_LOCK

Lock bit for region0 ctrl_aes_mode

CTRL_AES_MODE_R1

AES mode of region1

CTRL_AES_MODE_R1_LOCK

Lock bit for region1 ctrl_aes_mode

CTRL_CLK_EN

Clock enable input, low inactive

CTRL_CLK_EN_LOCK

Lock bit for ctrl_clk_en

CTRL_SFTRST_N

Soft reset input, low active

CTRL_SFTRST_N_LOCK

Lock bit for ctrl_sftrst

KEY_REGION_SEL

AES key region select

KEY_REGION_SEL_LOCK

Lock bit for key_region_sel

KEY_VALID

AES-128 key is ready

KEY_VALID_LOCK

Lock bit for key_valid

LITTLE_ENDIAN

Endian swap control for the 16 bytes input and output data of AES core.

LITTLE_ENDIAN_LOCK

Lock bit for little_endian

REGION0_KEY_LOCK

Lock bit for region0 AES key

REGION1_ADDR_LOCK

Lock bit for region1 address boundary

REGION1_KEY_LOCK

Lock bit for region1 AES key

SECURITY_LEVEL_R0

Security level of the allowed access for memory region0

SECURITY_LEVEL_R0_LOCK

Lock bits for security_level_r0

SECURITY_LEVEL_R1

Security level of the allowed access for memory region1

SECURITY_LEVEL_R1_LOCK

Lock bits for security_level_r1