imxrt_ral::can3
Module CS42

Message Buffer 42 CS Register

Modules

BRS

Bit Rate Switch. This bit defines whether the bit rate is switched inside a CAN FD format frame.

CODE

Message Buffer Code. This 4-bit field can be accessed (read or write) by the CPU and by the FlexCAN module itself, as part of the message buffer matching and arbitration process.

DLC

Length of the data to be stored/transmitted.

EDL

Extended Data Length. This bit distinguishes between CAN format and CAN FD format frames. The EDL bit must not be set for Message Buffers configured to RANSWER with code field 0b1010.

ESI

Error State Indicator. This bit indicates if the transmitting node is error active or error passive.

IDE

ID Extended. One/zero for extended/standard format frame.

RTR

Remote Transmission Request. One/zero for remote/data frame.

SRR

Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP

Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.