Module imxrt_ral::ccm::CLPCR

CCM Low Power Control Register

Modules

ARM_CLK_DIS_ON_LPM

Define if ARM clocks (arm_clk, soc_mxclk, soc_pclk, soc_dbg_pclk, vl_wrck) will be disabled on wait mode

BYPASS_LPM_HS0

Bypass low power mode handshake. This bit should always be set to 1’b1 by software.

BYPASS_LPM_HS1

Bypass low power mode handshake. This bit should always be set to 1’b1 by software.

COSC_PWRDOWN

In run mode, software can manually control powering down of on chip oscillator, i

DIS_REF_OSC

dis_ref_osc - in run mode, software can manually control closing of external reference oscillator clock, i

LPM

Setting the low power mode that system will enter on next assertion of dsm_request signal.

MASK_CORE0_WFI

Mask WFI of core0 for entering low power mode Assertion of all bits[27:22] will generate low power mode request

MASK_L2CC_IDLE

Mask L2CC IDLE for entering low power mode

MASK_SCU_IDLE

Mask SCU IDLE for entering low power mode Assertion of all bits[27:22] will generate low power mode request

SBYOS

Standby clock oscillator bit

STBY_COUNT

Standby counter definition

VSTBY

Voltage standby request bit