imxrt_ral::csi
Module CSICR18

CSI Control Register 18

Modules

AHB_HPROT

Hprot value in AHB bus protocol.

BASEADDR_CHANGE_ERROR_IE

Base address change error interrupt enable signal.

BASEADDR_SWITCH_EN

When this bit is enabled, CSI DMA will switch the base address according to BASEADDR_SWITCH_SEL rather than atomically by DMA completed

BASEADDR_SWITCH_SEL

CSI 2 base addresses switching method. When using this bit, BASEADDR_SWITCH_EN is 1.

CSI_ENABLE

CSI global enable signal

DEINTERLACE_EN

This bit is used to select the output method When input is standard CCIR656 video.

DMA_FIELD1_DONE_IE

When in interlace mode, field 1 done interrupt enable.

FIELD0_DONE_IE

In interlace mode, fileld 0 means interrupt enabled.

LAST_DMA_REQ_SEL

Choosing the last DMA request condition.

MASK_OPTION

These bits used to choose the method to mask the CSI input.

PARALLEL24_EN

When input is parallel rgb888/yuv444 24bit, this bit can be enabled.

RGB888A_FORMAT_SEL

Output is 32-bit format.