imxrt_ral::dcdc
Module REG1

DCDC Register 1

Modules

LOOPCTRL_EN_HYST

Enable hysteresis in switching converter common mode analog comparators

LOOPCTRL_HST_THRESH

increase the threshold detection for common mode analog comparator

LP_CMP_ISRC_SEL

set the current bias of low power comparator 0x0: 50 nA 0x1: 100 nA 0x2: 200 nA 0x3: 400 nA

REG_FBK_SEL

select the feedback point of the internal regulator

REG_RLOAD_SW

control the load resistor of the internal regulator of DCDC, the load resistor is connected as default “1”, and need set to “0” to disconnect the load resistor

VBG_TRIM

trim bandgap voltage