Module imxrt_ral::dcp::CTRL_SET

DCP control register 0

Modules

CHANNEL_INTERRUPT_ENABLE

Per-channel interrupt enable bit

CLKGATE

This bit must be set to zero for a normal operation

ENABLE_CONTEXT_CACHING

The software must set this bit to enable the caching of contexts between the operations

ENABLE_CONTEXT_SWITCHING

Enable automatic context switching for the channels

GATHER_RESIDUAL_WRITES

The software must set this bit to enable the ragged writes to the unaligned buffers to be gathered between multiple write operations

PRESENT_CRYPTO

Indicates whether the crypto (cipher/hash) functions are present.

PRESENT_SHA

Indicates whether the SHA1/SHA2 functions are present.

SFTRST

Set this bit to zero to enable a normal DCP operation