Module imxrt_ral::dcp::STAT_CLR

DCP status register

Modules

CUR_CHANNEL

Current (active) channel (encoded)

IRQ

Indicates which channels have pending interrupt requests

OTP_KEY_READY

When set, it indicates that the OTP key is shifted from the fuse block and is ready for use.

READY_CHANNELS

Indicates which channels are ready to proceed with a transfer (the active channel is also included)