imxrt_ral::enet::TCSR0::TMODE
Module RW

Constants

TMODE_0

Timer Channel is disabled.

TMODE_1

Timer Channel is configured for Input Capture on rising edge.

TMODE_2

Timer Channel is configured for Input Capture on falling edge.

TMODE_3

Timer Channel is configured for Input Capture on both edges.

TMODE_4

Timer Channel is configured for Output Compare - software only.

TMODE_5

Timer Channel is configured for Output Compare - toggle output on compare.

TMODE_6

Timer Channel is configured for Output Compare - clear output on compare.

TMODE_7

Timer Channel is configured for Output Compare - set output on compare.

TMODE_9

Timer Channel is configured for Output Compare - set output on compare, clear output on overflow.

TMODE_10

Timer Channel is configured for Output Compare - clear output on compare, set output on overflow.

TMODE_14

Timer Channel is configured for Output Compare - pulse output low on compare for 1 to 32 1588-clock cycles as specified by TPWC.

TMODE_15

Timer Channel is configured for Output Compare - pulse output high on compare for 1 to 32 1588-clock cycles as specified by TPWC.