Module imxrt_ral::flexspi::FLSHCR2

Flash Control Register 2

Modules

ARDSEQID

Sequence Index for AHB Read triggered Command in LUT.

ARDSEQNUM

Sequence Number for AHB Read triggered Command in LUT.

AWRSEQID

Sequence Index for AHB Write triggered Command.

AWRSEQNUM

Sequence Number for AHB Write triggered Command.

AWRWAIT

For certain devices (such as FPGA), it need some time to write data into internal memory after the command sequences finished on FlexSPI interface

AWRWAITUNIT

AWRWAIT unit

CLRINSTRPTR

Clear the instruction pointer which is internally saved pointer by JMP_ON_CS. Refer Programmable Sequence Engine for details.