Module imxrt_ral::flexspi::STS2

Status Register 2

Modules

AREFLOCK

Flash A sample clock reference delay line locked.

AREFSEL

Flash A sample clock reference delay line delay cell number selection.

ASLVLOCK

Flash A sample clock slave delay line locked.

ASLVSEL

Flash A sample clock slave delay line delay cell number selection .

BREFLOCK

Flash B sample clock reference delay line locked.

BREFSEL

Flash B sample clock reference delay line delay cell number selection.

BSLVLOCK

Flash B sample clock slave delay line locked.

BSLVSEL

Flash B sample clock slave delay line delay cell number selection.