imxrt_ral::gpt
Module SR

GPT Status Register

Modules

IF1

See IF2

IF2

IF2 Input capture 2 Flag IF1 Input capture 1 Flag The IFn bit indicates that a capture event has occurred on Input Capture channel n

OF1

See OF3

OF2

See OF3

OF3

OF3 Output Compare 3 Flag OF2 Output Compare 2 Flag OF1 Output Compare 1 Flag The OFn bit indicates that a compare event has occurred on Output Compare channel n

ROV

Rollover Flag