Module imxrt_ral::imxrt101::imxrt1011::dma0

DMA

Modules

CDNE

Clear DONE Status Bit Register

CEEI

Clear Enable Error Interrupt Register

CERQ

Clear Enable Request Register

CERR

Clear Error Register

CINT

Clear Interrupt Request Register

CR

Control Register

DCHPRI0

Channel Priority Register

DCHPRI1

Channel Priority Register

DCHPRI2

Channel Priority Register

DCHPRI3

Channel Priority Register

DCHPRI4

Channel Priority Register

DCHPRI5

Channel Priority Register

DCHPRI6

Channel Priority Register

DCHPRI7

Channel Priority Register

DCHPRI8

Channel Priority Register

DCHPRI9

Channel Priority Register

DCHPRI10

Channel Priority Register

DCHPRI11

Channel Priority Register

DCHPRI12

Channel Priority Register

DCHPRI13

Channel Priority Register

DCHPRI14

Channel Priority Register

DCHPRI15

Channel Priority Register

EARS

Enable Asynchronous Request in Stop Register

EEI

Enable Error Interrupt Register

ERQ

Enable Request Register

ERR

Error Register

ES

Error Status Register

HRS

Hardware Request Status Register

INT

Interrupt Request Register

SEEI

Set Enable Error Interrupt Register

SERQ

Set Enable Request Register

SSRT

Set START Bit Register

TCD_ATTR0

TCD Transfer Attributes

TCD_ATTR1

TCD Transfer Attributes

TCD_ATTR2

TCD Transfer Attributes

TCD_ATTR3

TCD Transfer Attributes

TCD_ATTR4

TCD Transfer Attributes

TCD_ATTR5

TCD Transfer Attributes

TCD_ATTR6

TCD Transfer Attributes

TCD_ATTR7

TCD Transfer Attributes

TCD_ATTR8

TCD Transfer Attributes

TCD_ATTR9

TCD Transfer Attributes

TCD_ATTR10

TCD Transfer Attributes

TCD_ATTR11

TCD Transfer Attributes

TCD_ATTR12

TCD Transfer Attributes

TCD_ATTR13

TCD Transfer Attributes

TCD_ATTR14

TCD Transfer Attributes

TCD_ATTR15

TCD Transfer Attributes

TCD_BITER_ELINK

TCD_BITER_ELINKNO0 and TCD_BITER_ELINKYES0 TCD_BITER_ELINKNO0: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES0: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO1

TCD_BITER_ELINKNO1 and TCD_BITER_ELINKYES1 TCD_BITER_ELINKNO1: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES1: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO2

TCD_BITER_ELINKNO2 and TCD_BITER_ELINKYES2 TCD_BITER_ELINKNO2: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES2: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO3

TCD_BITER_ELINKNO3 and TCD_BITER_ELINKYES3 TCD_BITER_ELINKNO3: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES3: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO4

TCD_BITER_ELINKNO4 and TCD_BITER_ELINKYES4 TCD_BITER_ELINKNO4: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES4: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO5

TCD_BITER_ELINKNO5 and TCD_BITER_ELINKYES5 TCD_BITER_ELINKNO5: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES5: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO6

TCD_BITER_ELINKNO6 and TCD_BITER_ELINKYES6 TCD_BITER_ELINKNO6: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES6: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO7

TCD_BITER_ELINKNO7 and TCD_BITER_ELINKYES7 TCD_BITER_ELINKNO7: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES7: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO8

TCD_BITER_ELINKNO8 and TCD_BITER_ELINKYES8 TCD_BITER_ELINKNO8: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES8: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO9

TCD_BITER_ELINKNO9 and TCD_BITER_ELINKYES9 TCD_BITER_ELINKNO9: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES9: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO10

TCD_BITER_ELINKNO10 and TCD_BITER_ELINKYES10 TCD_BITER_ELINKNO10: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES10: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO11

TCD_BITER_ELINKNO11 and TCD_BITER_ELINKYES11 TCD_BITER_ELINKNO11: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES11: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO12

TCD_BITER_ELINKNO12 and TCD_BITER_ELINKYES12 TCD_BITER_ELINKNO12: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES12: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO13

TCD_BITER_ELINKNO13 and TCD_BITER_ELINKYES13 TCD_BITER_ELINKNO13: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES13: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO14

TCD_BITER_ELINKNO14 and TCD_BITER_ELINKYES14 TCD_BITER_ELINKNO14: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES14: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO15

TCD_BITER_ELINKNO15 and TCD_BITER_ELINKYES15 TCD_BITER_ELINKNO15: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES15: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINK

TCD_CITER_ELINKNO0 and TCD_CITER_ELINKYES0 TCD_CITER_ELINKNO0: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES0: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO1

TCD_CITER_ELINKNO1 and TCD_CITER_ELINKYES1 TCD_CITER_ELINKNO1: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES1: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO2

TCD_CITER_ELINKNO2 and TCD_CITER_ELINKYES2 TCD_CITER_ELINKNO2: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES2: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO3

TCD_CITER_ELINKNO3 and TCD_CITER_ELINKYES3 TCD_CITER_ELINKNO3: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES3: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO4

TCD_CITER_ELINKNO4 and TCD_CITER_ELINKYES4 TCD_CITER_ELINKNO4: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES4: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO5

TCD_CITER_ELINKNO5 and TCD_CITER_ELINKYES5 TCD_CITER_ELINKNO5: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES5: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO6

TCD_CITER_ELINKNO6 and TCD_CITER_ELINKYES6 TCD_CITER_ELINKNO6: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES6: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO7

TCD_CITER_ELINKNO7 and TCD_CITER_ELINKYES7 TCD_CITER_ELINKNO7: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES7: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO8

TCD_CITER_ELINKNO8 and TCD_CITER_ELINKYES8 TCD_CITER_ELINKNO8: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES8: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO9

TCD_CITER_ELINKNO9 and TCD_CITER_ELINKYES9 TCD_CITER_ELINKNO9: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES9: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO10

TCD_CITER_ELINKNO10 and TCD_CITER_ELINKYES10 TCD_CITER_ELINKNO10: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES10: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO11

TCD_CITER_ELINKNO11 and TCD_CITER_ELINKYES11 TCD_CITER_ELINKNO11: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES11: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO12

TCD_CITER_ELINKNO12 and TCD_CITER_ELINKYES12 TCD_CITER_ELINKNO12: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES12: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO13

TCD_CITER_ELINKNO13 and TCD_CITER_ELINKYES13 TCD_CITER_ELINKNO13: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES13: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO14

TCD_CITER_ELINKNO14 and TCD_CITER_ELINKYES14 TCD_CITER_ELINKNO14: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES14: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO15

TCD_CITER_ELINKNO15 and TCD_CITER_ELINKYES15 TCD_CITER_ELINKNO15: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES15: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CSR0

TCD Control and Status

TCD_CSR1

TCD Control and Status

TCD_CSR2

TCD Control and Status

TCD_CSR3

TCD Control and Status

TCD_CSR4

TCD Control and Status

TCD_CSR5

TCD Control and Status

TCD_CSR6

TCD Control and Status

TCD_CSR7

TCD Control and Status

TCD_CSR8

TCD Control and Status

TCD_CSR9

TCD Control and Status

TCD_CSR10

TCD Control and Status

TCD_CSR11

TCD Control and Status

TCD_CSR12

TCD Control and Status

TCD_CSR13

TCD Control and Status

TCD_CSR14

TCD Control and Status

TCD_CSR15

TCD Control and Status

TCD_DADDR0

TCD Destination Address

TCD_DADDR1

TCD Destination Address

TCD_DADDR2

TCD Destination Address

TCD_DADDR3

TCD Destination Address

TCD_DADDR4

TCD Destination Address

TCD_DADDR5

TCD Destination Address

TCD_DADDR6

TCD Destination Address

TCD_DADDR7

TCD Destination Address

TCD_DADDR8

TCD Destination Address

TCD_DADDR9

TCD Destination Address

TCD_DADDR10

TCD Destination Address

TCD_DADDR11

TCD Destination Address

TCD_DADDR12

TCD Destination Address

TCD_DADDR13

TCD Destination Address

TCD_DADDR14

TCD Destination Address

TCD_DADDR15

TCD Destination Address

TCD_DLASTSGA0

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA1

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA2

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA3

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA4

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA5

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA6

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA7

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA8

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA9

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA10

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA11

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA12

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA13

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA14

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA15

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DOFF0

TCD Signed Destination Address Offset

TCD_DOFF1

TCD Signed Destination Address Offset

TCD_DOFF2

TCD Signed Destination Address Offset

TCD_DOFF3

TCD Signed Destination Address Offset

TCD_DOFF4

TCD Signed Destination Address Offset

TCD_DOFF5

TCD Signed Destination Address Offset

TCD_DOFF6

TCD Signed Destination Address Offset

TCD_DOFF7

TCD Signed Destination Address Offset

TCD_DOFF8

TCD Signed Destination Address Offset

TCD_DOFF9

TCD Signed Destination Address Offset

TCD_DOFF10

TCD Signed Destination Address Offset

TCD_DOFF11

TCD Signed Destination Address Offset

TCD_DOFF12

TCD Signed Destination Address Offset

TCD_DOFF13

TCD Signed Destination Address Offset

TCD_DOFF14

TCD Signed Destination Address Offset

TCD_DOFF15

TCD Signed Destination Address Offset

TCD_NBYTES_ML

TCD_NBYTES_ML and TCD_NBYTES_MLOFFYES0 TCD_NBYTES_ML: TCD_NBYTES_MLNO0 and TCD_NBYTES_MLOFFNO0 TCD_NBYTES_MLNO0: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO0: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES0: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO1

TCD_NBYTES_MLNO1 and TCD_NBYTES_MLOFFYES1 TCD_NBYTES_MLNO1: TCD_NBYTES_MLNO1 and TCD_NBYTES_MLOFFNO1 TCD_NBYTES_MLNO1: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO1: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES1: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO2

TCD_NBYTES_MLNO2 and TCD_NBYTES_MLOFFYES2 TCD_NBYTES_MLNO2: TCD_NBYTES_MLNO2 and TCD_NBYTES_MLOFFNO2 TCD_NBYTES_MLNO2: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO2: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES2: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO3

TCD_NBYTES_MLNO3 and TCD_NBYTES_MLOFFYES3 TCD_NBYTES_MLNO3: TCD_NBYTES_MLNO3 and TCD_NBYTES_MLOFFNO3 TCD_NBYTES_MLNO3: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO3: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES3: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO4

TCD_NBYTES_MLNO4 and TCD_NBYTES_MLOFFYES4 TCD_NBYTES_MLNO4: TCD_NBYTES_MLNO4 and TCD_NBYTES_MLOFFNO4 TCD_NBYTES_MLNO4: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO4: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES4: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO5

TCD_NBYTES_MLNO5 and TCD_NBYTES_MLOFFYES5 TCD_NBYTES_MLNO5: TCD_NBYTES_MLNO5 and TCD_NBYTES_MLOFFNO5 TCD_NBYTES_MLNO5: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO5: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES5: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO6

TCD_NBYTES_MLNO6 and TCD_NBYTES_MLOFFYES6 TCD_NBYTES_MLNO6: TCD_NBYTES_MLNO6 and TCD_NBYTES_MLOFFNO6 TCD_NBYTES_MLNO6: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO6: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES6: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO7

TCD_NBYTES_MLNO7 and TCD_NBYTES_MLOFFYES7 TCD_NBYTES_MLNO7: TCD_NBYTES_MLNO7 and TCD_NBYTES_MLOFFNO7 TCD_NBYTES_MLNO7: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO7: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES7: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO8

TCD_NBYTES_MLNO8 and TCD_NBYTES_MLOFFYES8 TCD_NBYTES_MLNO8: TCD_NBYTES_MLNO8 and TCD_NBYTES_MLOFFNO8 TCD_NBYTES_MLNO8: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO8: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES8: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO9

TCD_NBYTES_MLNO9 and TCD_NBYTES_MLOFFYES9 TCD_NBYTES_MLNO9: TCD_NBYTES_MLNO9 and TCD_NBYTES_MLOFFNO9 TCD_NBYTES_MLNO9: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO9: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES9: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO10

TCD_NBYTES_MLNO10 and TCD_NBYTES_MLOFFYES10 TCD_NBYTES_MLNO10: TCD_NBYTES_MLNO10 and TCD_NBYTES_MLOFFNO10 TCD_NBYTES_MLNO10: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO10: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES10: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO11

TCD_NBYTES_MLNO11 and TCD_NBYTES_MLOFFYES11 TCD_NBYTES_MLNO11: TCD_NBYTES_MLNO11 and TCD_NBYTES_MLOFFNO11 TCD_NBYTES_MLNO11: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO11: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES11: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO12

TCD_NBYTES_MLNO12 and TCD_NBYTES_MLOFFYES12 TCD_NBYTES_MLNO12: TCD_NBYTES_MLNO12 and TCD_NBYTES_MLOFFNO12 TCD_NBYTES_MLNO12: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO12: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES12: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO13

TCD_NBYTES_MLNO13 and TCD_NBYTES_MLOFFYES13 TCD_NBYTES_MLNO13: TCD_NBYTES_MLNO13 and TCD_NBYTES_MLOFFNO13 TCD_NBYTES_MLNO13: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO13: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES13: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO14

TCD_NBYTES_MLNO14 and TCD_NBYTES_MLOFFYES14 TCD_NBYTES_MLNO14: TCD_NBYTES_MLNO14 and TCD_NBYTES_MLOFFNO14 TCD_NBYTES_MLNO14: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO14: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES14: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO15

TCD_NBYTES_MLNO15 and TCD_NBYTES_MLOFFYES15 TCD_NBYTES_MLNO15: TCD_NBYTES_MLNO15 and TCD_NBYTES_MLOFFNO15 TCD_NBYTES_MLNO15: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO15: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES15: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_SADDR0

TCD Source Address

TCD_SADDR1

TCD Source Address

TCD_SADDR2

TCD Source Address

TCD_SADDR3

TCD Source Address

TCD_SADDR4

TCD Source Address

TCD_SADDR5

TCD Source Address

TCD_SADDR6

TCD Source Address

TCD_SADDR7

TCD Source Address

TCD_SADDR8

TCD Source Address

TCD_SADDR9

TCD Source Address

TCD_SADDR10

TCD Source Address

TCD_SADDR11

TCD Source Address

TCD_SADDR12

TCD Source Address

TCD_SADDR13

TCD Source Address

TCD_SADDR14

TCD Source Address

TCD_SADDR15

TCD Source Address

TCD_SLAST0

TCD Last Source Address Adjustment

TCD_SLAST1

TCD Last Source Address Adjustment

TCD_SLAST2

TCD Last Source Address Adjustment

TCD_SLAST3

TCD Last Source Address Adjustment

TCD_SLAST4

TCD Last Source Address Adjustment

TCD_SLAST5

TCD Last Source Address Adjustment

TCD_SLAST6

TCD Last Source Address Adjustment

TCD_SLAST7

TCD Last Source Address Adjustment

TCD_SLAST8

TCD Last Source Address Adjustment

TCD_SLAST9

TCD Last Source Address Adjustment

TCD_SLAST10

TCD Last Source Address Adjustment

TCD_SLAST11

TCD Last Source Address Adjustment

TCD_SLAST12

TCD Last Source Address Adjustment

TCD_SLAST13

TCD Last Source Address Adjustment

TCD_SLAST14

TCD Last Source Address Adjustment

TCD_SLAST15

TCD Last Source Address Adjustment

TCD_SOFF0

TCD Signed Source Address Offset

TCD_SOFF1

TCD Signed Source Address Offset

TCD_SOFF2

TCD Signed Source Address Offset

TCD_SOFF3

TCD Signed Source Address Offset

TCD_SOFF4

TCD Signed Source Address Offset

TCD_SOFF5

TCD Signed Source Address Offset

TCD_SOFF6

TCD Signed Source Address Offset

TCD_SOFF7

TCD Signed Source Address Offset

TCD_SOFF8

TCD Signed Source Address Offset

TCD_SOFF9

TCD Signed Source Address Offset

TCD_SOFF10

TCD Signed Source Address Offset

TCD_SOFF11

TCD Signed Source Address Offset

TCD_SOFF12

TCD Signed Source Address Offset

TCD_SOFF13

TCD Signed Source Address Offset

TCD_SOFF14

TCD Signed Source Address Offset

TCD_SOFF15

TCD Signed Source Address Offset

Structs

DMA0

The DMA0 peripheral instance.

Instance

RegisterBlock

ResetValues

Constants

DMA0

Raw pointer to DMA0