Module imxrt_ral::imxrt101::imxrt1011::flexspi::INTEN

Interrupt Enable Register

Modules

AHBBUSERROREN

AHB Bus error interrupt enable.Refer Interrupts chapter for more details.

AHBCMDERREN

AHB triggered Command Sequences Error Detected interrupt enable.

AHBCMDGEEN

AHB triggered Command Sequences Grant Timeout interrupt enable.

IPCMDDONEEN

IP triggered Command Sequences Execution finished interrupt enable.

IPCMDERREN

IP triggered Command Sequences Error Detected interrupt enable.

IPCMDGEEN

IP triggered Command Sequences Grant Timeout interrupt enable.

IPRXWAEN

IP RX FIFO WaterMark available interrupt enable.

IPTXWEEN

IP TX FIFO WaterMark empty interrupt enable.

KEYDONEEN

OTFAD key blob processing done interrupt enable.Refer Interrupts chapter for more details.

KEYERROREN

OTFAD key blob processing error interrupt enable.Refer Interrupts chapter for more details.

SCKSTOPBYRDEN

SCLK is stopped during command sequence because Async RX FIFO full interrupt enable.

SCKSTOPBYWREN

SCLK is stopped during command sequence because Async TX FIFO empty interrupt enable.

SEQTIMEOUTEN

Sequence execution timeout interrupt enable.Refer Interrupts chapter for more details.