Module imxrt_ral::imxrt101::imxrt1015::tmr1::CTRL0::CM::RW

Read-write values

Constants

CM_0
0b000: No operation

CM_1
0b001: Count rising edges of primary sourceRising edges are counted only when SCTRL[IPS] = 0. Falling edges are counted when SCTRL[IPS] = 1. If the primary count source is IP bus clock divide by 1, only rising edges are counted regardless of the value of SCTRL[IPS].

CM_2
0b010: Count rising and falling edges of primary sourceIP bus clock divide by 1 cannot be used as a primary count source in edge count mode.

CM_3
0b011: Count rising edges of primary source while secondary input high active

CM_4
0b100: Quadrature count mode, uses primary and secondary sources

CM_5
0b101: Count rising edges of primary source; secondary source specifies directionRising edges are counted only when SCTRL[IPS] = 0. Falling edges are counted when SCTRL[IPS] = 1.

CM_6
0b110: Edge of secondary source triggers primary count until compare

CM_7
0b111: Cascaded counter mode (up/down)The primary count source must be set to one of the counter outputs.