Module imxrt_ral::imxrt101::peripherals::csu::CSL0

Config security level register

Modules

LOCK_S1
The lock bit corresponding to the first slave. It is written by the secure software.

LOCK_S2
The lock bit corresponding to the second slave. It is written by the secure software.

NSR_S1
Non-secure supervisor read access control for the first slave

NSR_S2
Non-secure supervisor read access control for the second slave

NSW_S1
Non-secure supervisor write access control for the first slave

NSW_S2
Non-secure supervisor write access control for the second slave

NUR_S1
Non-secure user read access control for the first slave

NUR_S2
Non-secure user read access control for the second slave

NUW_S1
Non-secure user write access control for the first slave

NUW_S2
Non-secure user write access control for the second slave

SSR_S1
Secure supervisor read access control for the first slave

SSR_S2
Secure supervisor read access control for the second slave

SSW_S1
Secure supervisor write access control for the first slave

SSW_S2
Secure supervisor write access control for the second slave

SUR_S1
Secure user read access control for the first slave

SUR_S2
Secure user read access control for the second slave

SUW_S1
Secure user write access control for the first slave

SUW_S2
Secure user write access control for the second slave