Module imxrt_ral::imxrt101::peripherals::csu::HP0

HP0 register

Modules

HP_CSI

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the CSI

HP_DCP

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the DCP

HP_DMA

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the eDMA

HP_ENET

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the ENET

HP_LCDIF

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the LCDIF

HP_PXP

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the PXP

HP_TPSMP

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the TPSMP

HP_USB

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the USB

HP_USDHC1

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the USDHC1

HP_USDHC2

Determines whether the register value of the corresponding HP field is passed as the hprot[1] of the USDHC2

L_CSI

Lock bit set by the TZ software for the CSI

L_DCP

Lock bit set by the TZ software for the DCP

L_DMA

Lock bit set by the TZ software for the eDMA

L_ENET

Lock bit set by the TZ software for the ENET

L_LCDIF

Lock bit set by the TZ software for the LCDIF

L_PXP

Lock bit set by the TZ software for the PXP

L_TPSMP

Lock bit set by the TZ software for the TPSMP

L_USB

Lock bit set by the TZ software for the USB

L_USDHC1

Lock bit set by the TZ software for the USDHC1

L_USDHC2

Lock bit set by the TZ software for the USDHC2