Module imxrt_ral::imxrt102::imxrt1021::enet::TCSR0::TMODE::RW

Read-write values

Constants

TMODE_0
0b0000: Timer Channel is disabled.

TMODE_1
0b0001: Timer Channel is configured for Input Capture on rising edge.

TMODE_2
0b0010: Timer Channel is configured for Input Capture on falling edge.

TMODE_3
0b0011: Timer Channel is configured for Input Capture on both edges.

TMODE_4
0b0100: Timer Channel is configured for Output Compare - software only.

TMODE_5
0b0101: Timer Channel is configured for Output Compare - toggle output on compare.

TMODE_6
0b0110: Timer Channel is configured for Output Compare - clear output on compare.

TMODE_7
0b0111: Timer Channel is configured for Output Compare - set output on compare.

TMODE_9
0b0000: Timer Channel is configured for Output Compare - set output on compare, clear output on overflow.

TMODE_10
0b1010: Timer Channel is configured for Output Compare - clear output on compare, set output on overflow.

TMODE_14
0b1110: Timer Channel is configured for Output Compare - pulse output low on compare for 1 to 32 1588-clock cycles as specified by TPWC.

TMODE_15
0b1111: Timer Channel is configured for Output Compare - pulse output high on compare for 1 to 32 1588-clock cycles as specified by TPWC.