Module imxrt_ral::imxrt102::imxrt1021::flexio1::TIMCFG0::TIMRST::RW

Read-write values

Constants

TIMRST_0
0b000: Timer never reset

TIMRST_2
0b010: Timer reset on Timer Pin equal to Timer Output

TIMRST_3
0b011: Timer reset on Timer Trigger equal to Timer Output

TIMRST_4
0b100: Timer reset on Timer Pin rising edge

TIMRST_6
0b110: Timer reset on Trigger rising edge

TIMRST_7
0b111: Timer reset on Trigger rising or falling edge