Module imxrt_ral::imxrt102::imxrt1021::snvs::HPCR

SNVS_HP Control Register

Modules

BTN_CONFIG
Button Configuration

BTN_MASK
Button interrupt mask

DIS_PI
Disable periodic interrupt in the functional interrupt

HPCALB_EN
HP Real Time Counter Calibration Enabled Indicates that the time calibration mechanism is enabled.

HPCALB_VAL
HP Calibration Value Defines signed calibration value for the HP Real Time Counter

HPTA_EN
HP Time Alarm Enable When set, the time alarm interrupt is generated if the value in the HP Time Alarm Registers is equal to the value of the HP Real Time Counter

HP_TS
HP Time Synchronize

PI_EN
HP Periodic Interrupt Enable The periodic interrupt can be generated only if the HP Real Time Counter is enabled

PI_FREQ
Periodic Interrupt Frequency Defines frequency of the periodic interrupt

RTC_EN
HP Real Time Counter Enable