Module imxrt_ral::imxrt102::imxrt1021::snvs::LPSR

SNVS_LP Status Register

Modules

EO
Emergency Off This bit is set when a power off is requested.

ESVD
External Security Violation Detected Indicates that a security violation is detected on one of the HP security violation ports

ET1D
External Tampering 1 Detected

LPNS
LP Section is Non-Secured Indicates that LP section was provisioned/programmed in the non-secure state

LPS
LP Section is Secured Indicates that the LP section is provisioned/programmed in the secure or trusted state

LPTA
LP Time Alarm

MCR
Monotonic Counter Rollover

PGD
Power Supply Glitch Detected 0 No power supply glitch. 1 Power supply glitch is detected.

SED
Scan Exit Detected

SPO
Set Power Off The SPO bit is set when the power button is pressed longer than the configured debounce time

SRTCR
Secure Real Time Counter Rollover