Module imxrt_ral::imxrt102::imxrt1021::systemcontrol

System Control Block

Modules

ACTLR
Auxiliary Control Register,

AIRCR
Application Interrupt and Reset Control Register

BFAR
BusFault Address Register

CCR
Configuration and Control Register

CCSIDR
Cache Size ID Register

CFSR
Configurable Fault Status Register

CLIDR
Cache Level ID register

CM7_ABFSR
Auxiliary Bus Fault Status Register

CM7_AHBPCR
AHBP Control Register

CM7_AHBSCR
AHB Slave Control Register

CM7_CACR
L1 Cache Control Register

CM7_DTCMCR
Data Tightly-Coupled Memory Control Register

CM7_ITCMCR
Instruction Tightly-Coupled Memory Control Register

CPACR
Coprocessor Access Control Register

CPUID
CPUID Base Register

CSSELR
Cache Size Selection Register

CTR
Cache Type register

DCCIMVAC
Data cache clean and invalidate by address to PoC

DCCISW
Data cache clean and invalidate by set/way

DCCMVAC
Data cache clean by address to PoC

DCCMVAU
Data cache by address to PoU

DCCSW
Data cache clean by set/way

DCIMVAC
Data cache invalidate by address to Point of Coherency (PoC)

DCISW
Data cache invalidate by set/way

DFSR
Debug Fault Status Register

HFSR
HardFault Status register

ICIALLU
Instruction cache invalidate all to Point of Unification (PoU)

ICIMVAU
Instruction cache invalidate by address to PoU

ICSR
Interrupt Control and State Register

ID_AFR0
Auxiliary Feature Register

ID_DFR0
Debug Feature Register

ID_ISAR0
Instruction Set Attributes Register 0

ID_ISAR1
Instruction Set Attributes Register 1

ID_ISAR2
Instruction Set Attributes Register 2

ID_ISAR3
Instruction Set Attributes Register 3

ID_ISAR4
Instruction Set Attributes Register 4

ID_MMFR0
Memory Model Feature Register 0

ID_MMFR1
Memory Model Feature Register 1

ID_MMFR2
Memory Model Feature Register 2

ID_MMFR3
Memory Model Feature Register 3

ID_PFR0
Processor Feature Register 0

ID_PFR1
Processor Feature Register 1

MMFAR
MemManage Fault Address Register

SCR
System Control Register

SHCSR
System Handler Control and State Register

SHPR1
System Handler Priority Register 1

SHPR2
System Handler Priority Register 2

SHPR3
System Handler Priority Register 3

STIR
Instruction cache invalidate all to Point of Unification (PoU)

SystemControl
Access functions for the SystemControl peripheral instance

VTOR
Vector Table Offset Register

Structs

Instance

RegisterBlock

ResetValues

Constants

SystemControl
Raw pointer to SystemControl