Module imxrt_ral::imxrt105::imxrt1052::lcdif::PIGEON_0_0

Panel Interface Signal Generator Register

Modules

EN

Enable pigeon Mode on this signal

INC_SEL

Event to incrment local counter

MASK_CNT

When the global counter selected through MASK_CNT_SEL matches value in this reg, pigeon local counter start ticking

MASK_CNT_SEL

select global counters as mask condition, use together with MASK_CNT

OFFSET

offset on pclk unit

POL

Polarity of signal output

STATE_MASK

state_mask = (FS|FB|FD|FE) and (LS|LB|LD|LE) , select any combination of scan states as reference point for local counter to start ticking