Module imxrt_ral::imxrt105::imxrt1052::lcdif::STAT

LCD Interface Status Register

Modules

DMA_REQ

Reflects the current state of the DMA Request line for the LCDIF

LFIFO_COUNT

Read only view of the current count in Latency buffer (LFIFO).

LFIFO_EMPTY

Read only view of the signals that indicates LCD LFIFO is empty.

LFIFO_FULL

Read only view of the signals that indicates LCD LFIFO is full.

PRESENT

0: LCDIF not present on this product 1: LCDIF is present.

TXFIFO_EMPTY

Read only view of the signals that indicates LCD TXFIFO is empty.

TXFIFO_FULL

Read only view of the signals that indicates LCD TXFIFO is full.