Struct imxrt_ral::imxrt105::imxrt1052::lpi2c::LPI2C3

pub struct LPI2C3 { /* private fields */ }

The LPI2C3 peripheral instance.

This is a new type only for documentation purposes. When compiling for a target, this is defined as

pub type LPI2C3 = Instance<3>;

Implementations

impl LPI2C3

Access functions for the LPI2C3 peripheral instance

pub const reset: ResetValues = ResetValues { VERID: 0x01000003, PARAM: 0x00000202, MCR: 0x00000000, MSR: 0x00000001, MIER: 0x00000000, MDER: 0x00000000, MCFGR0: 0x00000000, MCFGR1: 0x00000000, MCFGR2: 0x00000000, MCFGR3: 0x00000000, MDMR: 0x00000000, MCCR0: 0x00000000, MCCR1: 0x00000000, MFCR: 0x00000000, MFSR: 0x00000000, MTDR: 0x00000000, MRDR: 0x00004000, SCR: 0x00000000, SSR: 0x00000000, SIER: 0x00000000, SDER: 0x00000000, SCFGR1: 0x00000000, SCFGR2: 0x00000000, SAMR: 0x00000000, SASR: 0x00004000, STAR: 0x00000000, STDR: 0x00000000, SRDR: 0x00004000, }

Reset values for each field in LPI2C3

pub fn take() -> Option<Self>

Safe access to LPI2C3

This function returns Some(Instance) if this instance is not currently taken, and None if it is. This ensures that if you do get Some(Instance), you are ensured unique access to the peripheral and there cannot be data races (unless other code uses unsafe, of course). You can then pass the Instance around to other functions as required. When you’re done with it, you can call release(instance) to return it.

Instance itself dereferences to a RegisterBlock, which provides access to the peripheral’s registers.

pub fn release(_: Self)

Release exclusive access to LPI2C3

This function allows you to return an Instance so that it is available to take() again. This function will panic if you return a different Instance or if this instance is not already taken.

pub unsafe fn steal() -> Self

Unsafely steal LPI2C3

This function is similar to take() but forcibly takes the Instance, marking it as taken irregardless of its previous state.

impl LPI2C3

pub const INTERRUPTS: [Interrupt; 0] = []

The interrupts associated with LPI2C3

Note: the values are invalid for a documentation build.