Module imxrt_ral::imxrt105::peripherals::can::CS0

Message Buffer 0 CS Register

Modules

CODE

Reserved

DLC

Length of the data to be stored/transmitted.

IDE

ID Extended. One/zero for extended/standard format frame.

RTR

Remote Transmission Request. One/zero for remote/data frame.

SRR

Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP

Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.