Module imxrt_ral::imxrt105::peripherals::ccm::CCOSR

CCM Clock Output Source Register

Modules

CLKO1_DIV

Setting the divider of CCM_CLKO1

CLKO1_EN

Enable of CCM_CLKO1 clock

CLKO1_SEL

Selection of the clock to be generated on CCM_CLKO1

CLKO2_DIV

Setting the divider of CCM_CLKO2

CLKO2_EN

Enable of CCM_CLKO2 clock

CLKO2_SEL

Selection of the clock to be generated on CCM_CLKO2

CLK_OUT_SEL

CCM_CLKO1 output to reflect CCM_CLKO1 or CCM_CLKO2 clocks