Module imxrt_ral::imxrt105::peripherals::ccm::CIMR

CCM Interrupt Mask Register

Modules

ARM_PODF_LOADED
mask interrupt generation due to frequency change of arm_podf

MASK_AHB_PODF_LOADED
mask interrupt generation due to frequency change of ahb_podf

MASK_COSC_READY
mask interrupt generation due to on board oscillator ready

MASK_LRF_PLL
mask interrupt generation due to lrf of PLLs

MASK_PERIPH2_CLK_SEL_LOADED
mask interrupt generation due to update of periph2_clk_sel.

MASK_PERIPH_CLK_SEL_LOADED
mask interrupt generation due to update of periph_clk_sel.

MASK_SEMC_PODF_LOADED
mask interrupt generation due to frequency change of semc_podf