Module imxrt_ral::imxrt105::peripherals::dcp::CH0STAT

DCP channel 0 status register

Modules

ERROR_CODE

Indicates the additional error codes for some of the error conditions

ERROR_DST

This bit indicates that a bus error occurred when storing to the destination buffer

ERROR_PACKET

This bit indicates that a bus error occurred when reading the packet or payload, or when writing the status back to the packet payload

ERROR_PAGEFAULT

This bit indicates that a page fault occurred while converting a virtual address to a physical address

ERROR_SETUP

This bit indicates that the hardware detected an invalid programming configuration (such as a buffer length that is not a multiple of the natural data size for the operation)

ERROR_SRC

This bit indicates that a bus error occurred when reading from the source buffer

HASH_MISMATCH

This bit indicates that a hashing check operation mismatched for the control packets that enable the HASH_CHECK bit

TAG

Indicates the tag from the last completed packet in the command structure