Module imxrt_ral::imxrt105::peripherals::dcp::PACKET1

DCP work packet 1 status register

Modules

CHAIN

Reflects whether the next command pointer register must be loaded into the channel’s current descriptor pointer

CHAIN_CONTIGUOUS

Reflects whether the next packet’s address is located following this packet’s payload.

CHECK_HASH

Reflects whether the calculated hash value must be compared to the hash provided in the payload.

CIPHER_ENCRYPT

When the cipher block is enabled, this bit indicates whether the operation is encryption or decryption

CIPHER_INIT

Reflects whether the cipher block must load the initialization vector from the payload for this operation

CONSTANT_FILL

When this bit is set (MEMCOPY and BLIT modes only), the DCP simply fills the destination buffer with the value found in the source address field

DECR_SEMAPHORE

Reflects whether the channel’s semaphore must be decremented at the end of the current operation

ENABLE_BLIT

Reflects whether the DCP must perform a blit operation

ENABLE_CIPHER

Reflects whether the selected cipher function must be enabled for this operation.

ENABLE_HASH

Reflects whether the selected hashing function must be enabled for this operation.

ENABLE_MEMCOPY

Reflects whether the selected hashing function should be enabled for this operation.

HASH_INIT

Reflects whether the current hashing block is the initial block in the hashing operation, so the hash registers must be initialized before the operation

HASH_OUTPUT

When the hashing is enabled, this bit controls whether the input or output data is hashed.

HASH_TERM

Reflects whether the current hashing block is the final block in the hashing operation, so the hash padding must be applied by the hardware

INPUT_BYTESWAP

Reflects whether the DCP engine byteswaps the input data (big-endian data).

INPUT_WORDSWAP

Reflects whether the DCP engine wordswaps the input data (big-endian data).

INTERRUPT

Reflects whether the channel must issue an interrupt upon the completion of the packet.

KEY_BYTESWAP

Reflects whether the DCP engine swaps the key bytes (big-endian key).

KEY_WORDSWAP

Reflects whether the DCP engine swaps the key words (big-endian key).

OTP_KEY

Reflects whether a hardware-based key must be used

OUTPUT_BYTESWAP

Reflects whether the DCP engine byteswaps the output data (big-endian data).

OUTPUT_WORDSWAP

Reflects whether the DCP engine wordswaps the output data (big-endian data).

PAYLOAD_KEY

When set, it indicates the payload contains the key

TAG

Packet Tag

TEST_SEMA_IRQ

This bit is used to test the channel semaphore transition to 0. FOR TEST USE ONLY!