Module imxrt_ral::imxrt105::peripherals::dma0

DMA

Used by: imxrt1051, imxrt1052

Modules

CDNE

Clear DONE Status Bit Register

CEEI

Clear Enable Error Interrupt Register

CERQ

Clear Enable Request Register

CERR

Clear Error Register

CINT

Clear Interrupt Request Register

CR

Control Register

DCHPRI0

Channel n Priority Register

DCHPRI1

Channel n Priority Register

DCHPRI2

Channel n Priority Register

DCHPRI3

Channel n Priority Register

DCHPRI4

Channel n Priority Register

DCHPRI5

Channel n Priority Register

DCHPRI6

Channel n Priority Register

DCHPRI7

Channel n Priority Register

DCHPRI8

Channel n Priority Register

DCHPRI9

Channel n Priority Register

DCHPRI10

Channel n Priority Register

DCHPRI11

Channel n Priority Register

DCHPRI12

Channel n Priority Register

DCHPRI13

Channel n Priority Register

DCHPRI14

Channel n Priority Register

DCHPRI15

Channel n Priority Register

DCHPRI16

Channel n Priority Register

DCHPRI17

Channel n Priority Register

DCHPRI18

Channel n Priority Register

DCHPRI19

Channel n Priority Register

DCHPRI20

Channel n Priority Register

DCHPRI21

Channel n Priority Register

DCHPRI22

Channel n Priority Register

DCHPRI23

Channel n Priority Register

DCHPRI24

Channel n Priority Register

DCHPRI25

Channel n Priority Register

DCHPRI26

Channel n Priority Register

DCHPRI27

Channel n Priority Register

DCHPRI28

Channel n Priority Register

DCHPRI29

Channel n Priority Register

DCHPRI30

Channel n Priority Register

DCHPRI31

Channel n Priority Register

EARS

Enable Asynchronous Request in Stop Register

EEI

Enable Error Interrupt Register

ERQ

Enable Request Register

ERR

Error Register

ES

Error Status Register

HRS

Hardware Request Status Register

INT

Interrupt Request Register

SEEI

Set Enable Error Interrupt Register

SERQ

Set Enable Request Register

SSRT

Set START Bit Register

TCD_ATTR0

TCD Transfer Attributes

TCD_ATTR1

TCD Transfer Attributes

TCD_ATTR2

TCD Transfer Attributes

TCD_ATTR3

TCD Transfer Attributes

TCD_ATTR4

TCD Transfer Attributes

TCD_ATTR5

TCD Transfer Attributes

TCD_ATTR6

TCD Transfer Attributes

TCD_ATTR7

TCD Transfer Attributes

TCD_ATTR8

TCD Transfer Attributes

TCD_ATTR9

TCD Transfer Attributes

TCD_ATTR10

TCD Transfer Attributes

TCD_ATTR11

TCD Transfer Attributes

TCD_ATTR12

TCD Transfer Attributes

TCD_ATTR13

TCD Transfer Attributes

TCD_ATTR14

TCD Transfer Attributes

TCD_ATTR15

TCD Transfer Attributes

TCD_ATTR16

TCD Transfer Attributes

TCD_ATTR17

TCD Transfer Attributes

TCD_ATTR18

TCD Transfer Attributes

TCD_ATTR19

TCD Transfer Attributes

TCD_ATTR20

TCD Transfer Attributes

TCD_ATTR21

TCD Transfer Attributes

TCD_ATTR22

TCD Transfer Attributes

TCD_ATTR23

TCD Transfer Attributes

TCD_ATTR24

TCD Transfer Attributes

TCD_ATTR25

TCD Transfer Attributes

TCD_ATTR26

TCD Transfer Attributes

TCD_ATTR27

TCD Transfer Attributes

TCD_ATTR28

TCD Transfer Attributes

TCD_ATTR29

TCD Transfer Attributes

TCD_ATTR30

TCD Transfer Attributes

TCD_ATTR31

TCD Transfer Attributes

TCD_BITER_ELINK

TCD_BITER_ELINKNO0 and TCD_BITER_ELINKYES0 TCD_BITER_ELINKNO0: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES0: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO1

TCD_BITER_ELINKNO1 and TCD_BITER_ELINKYES1 TCD_BITER_ELINKNO1: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES1: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO2

TCD_BITER_ELINKNO2 and TCD_BITER_ELINKYES2 TCD_BITER_ELINKNO2: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES2: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO3

TCD_BITER_ELINKNO3 and TCD_BITER_ELINKYES3 TCD_BITER_ELINKNO3: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES3: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO4

TCD_BITER_ELINKNO4 and TCD_BITER_ELINKYES4 TCD_BITER_ELINKNO4: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES4: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO5

TCD_BITER_ELINKNO5 and TCD_BITER_ELINKYES5 TCD_BITER_ELINKNO5: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES5: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO6

TCD_BITER_ELINKNO6 and TCD_BITER_ELINKYES6 TCD_BITER_ELINKNO6: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES6: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO7

TCD_BITER_ELINKNO7 and TCD_BITER_ELINKYES7 TCD_BITER_ELINKNO7: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES7: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO8

TCD_BITER_ELINKNO8 and TCD_BITER_ELINKYES8 TCD_BITER_ELINKNO8: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES8: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO9

TCD_BITER_ELINKNO9 and TCD_BITER_ELINKYES9 TCD_BITER_ELINKNO9: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES9: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO10

TCD_BITER_ELINKNO10 and TCD_BITER_ELINKYES10 TCD_BITER_ELINKNO10: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES10: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO11

TCD_BITER_ELINKNO11 and TCD_BITER_ELINKYES11 TCD_BITER_ELINKNO11: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES11: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO12

TCD_BITER_ELINKNO12 and TCD_BITER_ELINKYES12 TCD_BITER_ELINKNO12: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES12: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO13

TCD_BITER_ELINKNO13 and TCD_BITER_ELINKYES13 TCD_BITER_ELINKNO13: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES13: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO14

TCD_BITER_ELINKNO14 and TCD_BITER_ELINKYES14 TCD_BITER_ELINKNO14: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES14: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO15

TCD_BITER_ELINKNO15 and TCD_BITER_ELINKYES15 TCD_BITER_ELINKNO15: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES15: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO16

TCD_BITER_ELINKNO16 and TCD_BITER_ELINKYES16 TCD_BITER_ELINKNO16: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES16: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO17

TCD_BITER_ELINKNO17 and TCD_BITER_ELINKYES17 TCD_BITER_ELINKNO17: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES17: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO18

TCD_BITER_ELINKNO18 and TCD_BITER_ELINKYES18 TCD_BITER_ELINKNO18: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES18: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO19

TCD_BITER_ELINKNO19 and TCD_BITER_ELINKYES19 TCD_BITER_ELINKNO19: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES19: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO20

TCD_BITER_ELINKNO20 and TCD_BITER_ELINKYES20 TCD_BITER_ELINKNO20: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES20: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO21

TCD_BITER_ELINKNO21 and TCD_BITER_ELINKYES21 TCD_BITER_ELINKNO21: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES21: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO22

TCD_BITER_ELINKNO22 and TCD_BITER_ELINKYES22 TCD_BITER_ELINKNO22: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES22: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO23

TCD_BITER_ELINKNO23 and TCD_BITER_ELINKYES23 TCD_BITER_ELINKNO23: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES23: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO24

TCD_BITER_ELINKNO24 and TCD_BITER_ELINKYES24 TCD_BITER_ELINKNO24: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES24: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO25

TCD_BITER_ELINKNO25 and TCD_BITER_ELINKYES25 TCD_BITER_ELINKNO25: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES25: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO26

TCD_BITER_ELINKNO26 and TCD_BITER_ELINKYES26 TCD_BITER_ELINKNO26: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES26: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO27

TCD_BITER_ELINKNO27 and TCD_BITER_ELINKYES27 TCD_BITER_ELINKNO27: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES27: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO28

TCD_BITER_ELINKNO28 and TCD_BITER_ELINKYES28 TCD_BITER_ELINKNO28: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES28: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO29

TCD_BITER_ELINKNO29 and TCD_BITER_ELINKYES29 TCD_BITER_ELINKNO29: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES29: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO30

TCD_BITER_ELINKNO30 and TCD_BITER_ELINKYES30 TCD_BITER_ELINKNO30: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES30: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_BITER_ELINKNO31

TCD_BITER_ELINKNO31 and TCD_BITER_ELINKYES31 TCD_BITER_ELINKNO31: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_BITER_ELINKYES31: TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINK

TCD_CITER_ELINKNO0 and TCD_CITER_ELINKYES0 TCD_CITER_ELINKNO0: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES0: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO1

TCD_CITER_ELINKNO1 and TCD_CITER_ELINKYES1 TCD_CITER_ELINKNO1: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES1: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO2

TCD_CITER_ELINKNO2 and TCD_CITER_ELINKYES2 TCD_CITER_ELINKNO2: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES2: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO3

TCD_CITER_ELINKNO3 and TCD_CITER_ELINKYES3 TCD_CITER_ELINKNO3: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES3: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO4

TCD_CITER_ELINKNO4 and TCD_CITER_ELINKYES4 TCD_CITER_ELINKNO4: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES4: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO5

TCD_CITER_ELINKNO5 and TCD_CITER_ELINKYES5 TCD_CITER_ELINKNO5: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES5: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO6

TCD_CITER_ELINKNO6 and TCD_CITER_ELINKYES6 TCD_CITER_ELINKNO6: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES6: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO7

TCD_CITER_ELINKNO7 and TCD_CITER_ELINKYES7 TCD_CITER_ELINKNO7: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES7: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO8

TCD_CITER_ELINKNO8 and TCD_CITER_ELINKYES8 TCD_CITER_ELINKNO8: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES8: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO9

TCD_CITER_ELINKNO9 and TCD_CITER_ELINKYES9 TCD_CITER_ELINKNO9: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES9: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO10

TCD_CITER_ELINKNO10 and TCD_CITER_ELINKYES10 TCD_CITER_ELINKNO10: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES10: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO11

TCD_CITER_ELINKNO11 and TCD_CITER_ELINKYES11 TCD_CITER_ELINKNO11: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES11: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO12

TCD_CITER_ELINKNO12 and TCD_CITER_ELINKYES12 TCD_CITER_ELINKNO12: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES12: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO13

TCD_CITER_ELINKNO13 and TCD_CITER_ELINKYES13 TCD_CITER_ELINKNO13: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES13: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO14

TCD_CITER_ELINKNO14 and TCD_CITER_ELINKYES14 TCD_CITER_ELINKNO14: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES14: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO15

TCD_CITER_ELINKNO15 and TCD_CITER_ELINKYES15 TCD_CITER_ELINKNO15: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES15: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO16

TCD_CITER_ELINKNO16 and TCD_CITER_ELINKYES16 TCD_CITER_ELINKNO16: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES16: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO17

TCD_CITER_ELINKNO17 and TCD_CITER_ELINKYES17 TCD_CITER_ELINKNO17: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES17: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO18

TCD_CITER_ELINKNO18 and TCD_CITER_ELINKYES18 TCD_CITER_ELINKNO18: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES18: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO19

TCD_CITER_ELINKNO19 and TCD_CITER_ELINKYES19 TCD_CITER_ELINKNO19: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES19: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO20

TCD_CITER_ELINKNO20 and TCD_CITER_ELINKYES20 TCD_CITER_ELINKNO20: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES20: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO21

TCD_CITER_ELINKNO21 and TCD_CITER_ELINKYES21 TCD_CITER_ELINKNO21: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES21: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO22

TCD_CITER_ELINKNO22 and TCD_CITER_ELINKYES22 TCD_CITER_ELINKNO22: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES22: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO23

TCD_CITER_ELINKNO23 and TCD_CITER_ELINKYES23 TCD_CITER_ELINKNO23: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES23: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO24

TCD_CITER_ELINKNO24 and TCD_CITER_ELINKYES24 TCD_CITER_ELINKNO24: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES24: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO25

TCD_CITER_ELINKNO25 and TCD_CITER_ELINKYES25 TCD_CITER_ELINKNO25: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES25: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO26

TCD_CITER_ELINKNO26 and TCD_CITER_ELINKYES26 TCD_CITER_ELINKNO26: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES26: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO27

TCD_CITER_ELINKNO27 and TCD_CITER_ELINKYES27 TCD_CITER_ELINKNO27: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES27: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO28

TCD_CITER_ELINKNO28 and TCD_CITER_ELINKYES28 TCD_CITER_ELINKNO28: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES28: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO29

TCD_CITER_ELINKNO29 and TCD_CITER_ELINKYES29 TCD_CITER_ELINKNO29: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES29: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO30

TCD_CITER_ELINKNO30 and TCD_CITER_ELINKYES30 TCD_CITER_ELINKNO30: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES30: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CITER_ELINKNO31

TCD_CITER_ELINKNO31 and TCD_CITER_ELINKYES31 TCD_CITER_ELINKNO31: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled) TCD_CITER_ELINKYES31: TCD Current Minor Loop Link, Major Loop Count (Channel Linking Enabled)

TCD_CSR0

TCD Control and Status

TCD_CSR1

TCD Control and Status

TCD_CSR2

TCD Control and Status

TCD_CSR3

TCD Control and Status

TCD_CSR4

TCD Control and Status

TCD_CSR5

TCD Control and Status

TCD_CSR6

TCD Control and Status

TCD_CSR7

TCD Control and Status

TCD_CSR8

TCD Control and Status

TCD_CSR9

TCD Control and Status

TCD_CSR10

TCD Control and Status

TCD_CSR11

TCD Control and Status

TCD_CSR12

TCD Control and Status

TCD_CSR13

TCD Control and Status

TCD_CSR14

TCD Control and Status

TCD_CSR15

TCD Control and Status

TCD_CSR16

TCD Control and Status

TCD_CSR17

TCD Control and Status

TCD_CSR18

TCD Control and Status

TCD_CSR19

TCD Control and Status

TCD_CSR20

TCD Control and Status

TCD_CSR21

TCD Control and Status

TCD_CSR22

TCD Control and Status

TCD_CSR23

TCD Control and Status

TCD_CSR24

TCD Control and Status

TCD_CSR25

TCD Control and Status

TCD_CSR26

TCD Control and Status

TCD_CSR27

TCD Control and Status

TCD_CSR28

TCD Control and Status

TCD_CSR29

TCD Control and Status

TCD_CSR30

TCD Control and Status

TCD_CSR31

TCD Control and Status

TCD_DADDR0

TCD Destination Address

TCD_DADDR1

TCD Destination Address

TCD_DADDR2

TCD Destination Address

TCD_DADDR3

TCD Destination Address

TCD_DADDR4

TCD Destination Address

TCD_DADDR5

TCD Destination Address

TCD_DADDR6

TCD Destination Address

TCD_DADDR7

TCD Destination Address

TCD_DADDR8

TCD Destination Address

TCD_DADDR9

TCD Destination Address

TCD_DADDR10

TCD Destination Address

TCD_DADDR11

TCD Destination Address

TCD_DADDR12

TCD Destination Address

TCD_DADDR13

TCD Destination Address

TCD_DADDR14

TCD Destination Address

TCD_DADDR15

TCD Destination Address

TCD_DADDR16

TCD Destination Address

TCD_DADDR17

TCD Destination Address

TCD_DADDR18

TCD Destination Address

TCD_DADDR19

TCD Destination Address

TCD_DADDR20

TCD Destination Address

TCD_DADDR21

TCD Destination Address

TCD_DADDR22

TCD Destination Address

TCD_DADDR23

TCD Destination Address

TCD_DADDR24

TCD Destination Address

TCD_DADDR25

TCD Destination Address

TCD_DADDR26

TCD Destination Address

TCD_DADDR27

TCD Destination Address

TCD_DADDR28

TCD Destination Address

TCD_DADDR29

TCD Destination Address

TCD_DADDR30

TCD Destination Address

TCD_DADDR31

TCD Destination Address

TCD_DLASTSGA0

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA1

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA2

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA3

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA4

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA5

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA6

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA7

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA8

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA9

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA10

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA11

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA12

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA13

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA14

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA15

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA16

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA17

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA18

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA19

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA20

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA21

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA22

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA23

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA24

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA25

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA26

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA27

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA28

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA29

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA30

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DLASTSGA31

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DOFF0

TCD Signed Destination Address Offset

TCD_DOFF1

TCD Signed Destination Address Offset

TCD_DOFF2

TCD Signed Destination Address Offset

TCD_DOFF3

TCD Signed Destination Address Offset

TCD_DOFF4

TCD Signed Destination Address Offset

TCD_DOFF5

TCD Signed Destination Address Offset

TCD_DOFF6

TCD Signed Destination Address Offset

TCD_DOFF7

TCD Signed Destination Address Offset

TCD_DOFF8

TCD Signed Destination Address Offset

TCD_DOFF9

TCD Signed Destination Address Offset

TCD_DOFF10

TCD Signed Destination Address Offset

TCD_DOFF11

TCD Signed Destination Address Offset

TCD_DOFF12

TCD Signed Destination Address Offset

TCD_DOFF13

TCD Signed Destination Address Offset

TCD_DOFF14

TCD Signed Destination Address Offset

TCD_DOFF15

TCD Signed Destination Address Offset

TCD_DOFF16

TCD Signed Destination Address Offset

TCD_DOFF17

TCD Signed Destination Address Offset

TCD_DOFF18

TCD Signed Destination Address Offset

TCD_DOFF19

TCD Signed Destination Address Offset

TCD_DOFF20

TCD Signed Destination Address Offset

TCD_DOFF21

TCD Signed Destination Address Offset

TCD_DOFF22

TCD Signed Destination Address Offset

TCD_DOFF23

TCD Signed Destination Address Offset

TCD_DOFF24

TCD Signed Destination Address Offset

TCD_DOFF25

TCD Signed Destination Address Offset

TCD_DOFF26

TCD Signed Destination Address Offset

TCD_DOFF27

TCD Signed Destination Address Offset

TCD_DOFF28

TCD Signed Destination Address Offset

TCD_DOFF29

TCD Signed Destination Address Offset

TCD_DOFF30

TCD Signed Destination Address Offset

TCD_DOFF31

TCD Signed Destination Address Offset

TCD_NBYTES_ML

TCD_NBYTES_ML and TCD_NBYTES_MLOFFYES0 TCD_NBYTES_ML: TCD_NBYTES_MLNO0 and TCD_NBYTES_MLOFFNO0 TCD_NBYTES_MLNO0: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO0: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES0: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO1

TCD_NBYTES_MLNO1 and TCD_NBYTES_MLOFFYES1 TCD_NBYTES_MLNO1: TCD_NBYTES_MLNO1 and TCD_NBYTES_MLOFFNO1 TCD_NBYTES_MLNO1: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO1: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES1: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO2

TCD_NBYTES_MLNO2 and TCD_NBYTES_MLOFFYES2 TCD_NBYTES_MLNO2: TCD_NBYTES_MLNO2 and TCD_NBYTES_MLOFFNO2 TCD_NBYTES_MLNO2: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO2: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES2: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO3

TCD_NBYTES_MLNO3 and TCD_NBYTES_MLOFFYES3 TCD_NBYTES_MLNO3: TCD_NBYTES_MLNO3 and TCD_NBYTES_MLOFFNO3 TCD_NBYTES_MLNO3: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO3: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES3: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO4

TCD_NBYTES_MLNO4 and TCD_NBYTES_MLOFFYES4 TCD_NBYTES_MLNO4: TCD_NBYTES_MLNO4 and TCD_NBYTES_MLOFFNO4 TCD_NBYTES_MLNO4: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO4: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES4: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO5

TCD_NBYTES_MLNO5 and TCD_NBYTES_MLOFFYES5 TCD_NBYTES_MLNO5: TCD_NBYTES_MLNO5 and TCD_NBYTES_MLOFFNO5 TCD_NBYTES_MLNO5: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO5: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES5: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO6

TCD_NBYTES_MLNO6 and TCD_NBYTES_MLOFFYES6 TCD_NBYTES_MLNO6: TCD_NBYTES_MLNO6 and TCD_NBYTES_MLOFFNO6 TCD_NBYTES_MLNO6: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO6: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES6: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO7

TCD_NBYTES_MLNO7 and TCD_NBYTES_MLOFFYES7 TCD_NBYTES_MLNO7: TCD_NBYTES_MLNO7 and TCD_NBYTES_MLOFFNO7 TCD_NBYTES_MLNO7: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO7: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES7: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO8

TCD_NBYTES_MLNO8 and TCD_NBYTES_MLOFFYES8 TCD_NBYTES_MLNO8: TCD_NBYTES_MLNO8 and TCD_NBYTES_MLOFFNO8 TCD_NBYTES_MLNO8: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO8: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES8: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO9

TCD_NBYTES_MLNO9 and TCD_NBYTES_MLOFFYES9 TCD_NBYTES_MLNO9: TCD_NBYTES_MLNO9 and TCD_NBYTES_MLOFFNO9 TCD_NBYTES_MLNO9: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO9: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES9: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO10

TCD_NBYTES_MLNO10 and TCD_NBYTES_MLOFFYES10 TCD_NBYTES_MLNO10: TCD_NBYTES_MLNO10 and TCD_NBYTES_MLOFFNO10 TCD_NBYTES_MLNO10: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO10: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES10: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO11

TCD_NBYTES_MLNO11 and TCD_NBYTES_MLOFFYES11 TCD_NBYTES_MLNO11: TCD_NBYTES_MLNO11 and TCD_NBYTES_MLOFFNO11 TCD_NBYTES_MLNO11: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO11: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES11: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO12

TCD_NBYTES_MLNO12 and TCD_NBYTES_MLOFFYES12 TCD_NBYTES_MLNO12: TCD_NBYTES_MLNO12 and TCD_NBYTES_MLOFFNO12 TCD_NBYTES_MLNO12: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO12: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES12: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO13

TCD_NBYTES_MLNO13 and TCD_NBYTES_MLOFFYES13 TCD_NBYTES_MLNO13: TCD_NBYTES_MLNO13 and TCD_NBYTES_MLOFFNO13 TCD_NBYTES_MLNO13: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO13: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES13: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO14

TCD_NBYTES_MLNO14 and TCD_NBYTES_MLOFFYES14 TCD_NBYTES_MLNO14: TCD_NBYTES_MLNO14 and TCD_NBYTES_MLOFFNO14 TCD_NBYTES_MLNO14: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO14: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES14: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO15

TCD_NBYTES_MLNO15 and TCD_NBYTES_MLOFFYES15 TCD_NBYTES_MLNO15: TCD_NBYTES_MLNO15 and TCD_NBYTES_MLOFFNO15 TCD_NBYTES_MLNO15: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO15: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES15: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO16

TCD_NBYTES_MLNO16 and TCD_NBYTES_MLOFFYES16 TCD_NBYTES_MLNO16: TCD_NBYTES_MLNO16 and TCD_NBYTES_MLOFFNO16 TCD_NBYTES_MLNO16: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO16: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES16: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO17

TCD_NBYTES_MLNO17 and TCD_NBYTES_MLOFFYES17 TCD_NBYTES_MLNO17: TCD_NBYTES_MLNO17 and TCD_NBYTES_MLOFFNO17 TCD_NBYTES_MLNO17: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO17: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES17: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO18

TCD_NBYTES_MLNO18 and TCD_NBYTES_MLOFFYES18 TCD_NBYTES_MLNO18: TCD_NBYTES_MLNO18 and TCD_NBYTES_MLOFFNO18 TCD_NBYTES_MLNO18: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO18: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES18: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO19

TCD_NBYTES_MLNO19 and TCD_NBYTES_MLOFFYES19 TCD_NBYTES_MLNO19: TCD_NBYTES_MLNO19 and TCD_NBYTES_MLOFFNO19 TCD_NBYTES_MLNO19: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO19: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES19: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO20

TCD_NBYTES_MLNO20 and TCD_NBYTES_MLOFFYES20 TCD_NBYTES_MLNO20: TCD_NBYTES_MLNO20 and TCD_NBYTES_MLOFFNO20 TCD_NBYTES_MLNO20: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO20: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES20: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO21

TCD_NBYTES_MLNO21 and TCD_NBYTES_MLOFFYES21 TCD_NBYTES_MLNO21: TCD_NBYTES_MLNO21 and TCD_NBYTES_MLOFFNO21 TCD_NBYTES_MLNO21: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO21: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES21: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO22

TCD_NBYTES_MLNO22 and TCD_NBYTES_MLOFFYES22 TCD_NBYTES_MLNO22: TCD_NBYTES_MLNO22 and TCD_NBYTES_MLOFFNO22 TCD_NBYTES_MLNO22: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO22: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES22: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO23

TCD_NBYTES_MLNO23 and TCD_NBYTES_MLOFFYES23 TCD_NBYTES_MLNO23: TCD_NBYTES_MLNO23 and TCD_NBYTES_MLOFFNO23 TCD_NBYTES_MLNO23: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO23: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES23: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO24

TCD_NBYTES_MLNO24 and TCD_NBYTES_MLOFFYES24 TCD_NBYTES_MLNO24: TCD_NBYTES_MLNO24 and TCD_NBYTES_MLOFFNO24 TCD_NBYTES_MLNO24: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO24: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES24: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO25

TCD_NBYTES_MLNO25 and TCD_NBYTES_MLOFFYES25 TCD_NBYTES_MLNO25: TCD_NBYTES_MLNO25 and TCD_NBYTES_MLOFFNO25 TCD_NBYTES_MLNO25: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO25: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES25: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO26

TCD_NBYTES_MLNO26 and TCD_NBYTES_MLOFFYES26 TCD_NBYTES_MLNO26: TCD_NBYTES_MLNO26 and TCD_NBYTES_MLOFFNO26 TCD_NBYTES_MLNO26: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO26: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES26: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO27

TCD_NBYTES_MLNO27 and TCD_NBYTES_MLOFFYES27 TCD_NBYTES_MLNO27: TCD_NBYTES_MLNO27 and TCD_NBYTES_MLOFFNO27 TCD_NBYTES_MLNO27: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO27: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES27: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO28

TCD_NBYTES_MLNO28 and TCD_NBYTES_MLOFFYES28 TCD_NBYTES_MLNO28: TCD_NBYTES_MLNO28 and TCD_NBYTES_MLOFFNO28 TCD_NBYTES_MLNO28: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO28: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES28: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO29

TCD_NBYTES_MLNO29 and TCD_NBYTES_MLOFFYES29 TCD_NBYTES_MLNO29: TCD_NBYTES_MLNO29 and TCD_NBYTES_MLOFFNO29 TCD_NBYTES_MLNO29: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO29: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES29: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO30

TCD_NBYTES_MLNO30 and TCD_NBYTES_MLOFFYES30 TCD_NBYTES_MLNO30: TCD_NBYTES_MLNO30 and TCD_NBYTES_MLOFFNO30 TCD_NBYTES_MLNO30: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO30: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES30: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_NBYTES_MLNO31

TCD_NBYTES_MLNO31 and TCD_NBYTES_MLOFFYES31 TCD_NBYTES_MLNO31: TCD_NBYTES_MLNO31 and TCD_NBYTES_MLOFFNO31 TCD_NBYTES_MLNO31: TCD Minor Byte Count (Minor Loop Mapping Disabled) TCD_NBYTES_MLOFFNO31: TCD Signed Minor Loop Offset (Minor Loop Mapping Enabled and Offset Disabled) TCD_NBYTES_MLOFFYES31: TCD Signed Minor Loop Offset (Minor Loop Mapping and Offset Enabled)

TCD_SADDR0

TCD Source Address

TCD_SADDR1

TCD Source Address

TCD_SADDR2

TCD Source Address

TCD_SADDR3

TCD Source Address

TCD_SADDR4

TCD Source Address

TCD_SADDR5

TCD Source Address

TCD_SADDR6

TCD Source Address

TCD_SADDR7

TCD Source Address

TCD_SADDR8

TCD Source Address

TCD_SADDR9

TCD Source Address

TCD_SADDR10

TCD Source Address

TCD_SADDR11

TCD Source Address

TCD_SADDR12

TCD Source Address

TCD_SADDR13

TCD Source Address

TCD_SADDR14

TCD Source Address

TCD_SADDR15

TCD Source Address

TCD_SADDR16

TCD Source Address

TCD_SADDR17

TCD Source Address

TCD_SADDR18

TCD Source Address

TCD_SADDR19

TCD Source Address

TCD_SADDR20

TCD Source Address

TCD_SADDR21

TCD Source Address

TCD_SADDR22

TCD Source Address

TCD_SADDR23

TCD Source Address

TCD_SADDR24

TCD Source Address

TCD_SADDR25

TCD Source Address

TCD_SADDR26

TCD Source Address

TCD_SADDR27

TCD Source Address

TCD_SADDR28

TCD Source Address

TCD_SADDR29

TCD Source Address

TCD_SADDR30

TCD Source Address

TCD_SADDR31

TCD Source Address

TCD_SLAST0

TCD Last Source Address Adjustment

TCD_SLAST1

TCD Last Source Address Adjustment

TCD_SLAST2

TCD Last Source Address Adjustment

TCD_SLAST3

TCD Last Source Address Adjustment

TCD_SLAST4

TCD Last Source Address Adjustment

TCD_SLAST5

TCD Last Source Address Adjustment

TCD_SLAST6

TCD Last Source Address Adjustment

TCD_SLAST7

TCD Last Source Address Adjustment

TCD_SLAST8

TCD Last Source Address Adjustment

TCD_SLAST9

TCD Last Source Address Adjustment

TCD_SLAST10

TCD Last Source Address Adjustment

TCD_SLAST11

TCD Last Source Address Adjustment

TCD_SLAST12

TCD Last Source Address Adjustment

TCD_SLAST13

TCD Last Source Address Adjustment

TCD_SLAST14

TCD Last Source Address Adjustment

TCD_SLAST15

TCD Last Source Address Adjustment

TCD_SLAST16

TCD Last Source Address Adjustment

TCD_SLAST17

TCD Last Source Address Adjustment

TCD_SLAST18

TCD Last Source Address Adjustment

TCD_SLAST19

TCD Last Source Address Adjustment

TCD_SLAST20

TCD Last Source Address Adjustment

TCD_SLAST21

TCD Last Source Address Adjustment

TCD_SLAST22

TCD Last Source Address Adjustment

TCD_SLAST23

TCD Last Source Address Adjustment

TCD_SLAST24

TCD Last Source Address Adjustment

TCD_SLAST25

TCD Last Source Address Adjustment

TCD_SLAST26

TCD Last Source Address Adjustment

TCD_SLAST27

TCD Last Source Address Adjustment

TCD_SLAST28

TCD Last Source Address Adjustment

TCD_SLAST29

TCD Last Source Address Adjustment

TCD_SLAST30

TCD Last Source Address Adjustment

TCD_SLAST31

TCD Last Source Address Adjustment

TCD_SOFF0

TCD Signed Source Address Offset

TCD_SOFF1

TCD Signed Source Address Offset

TCD_SOFF2

TCD Signed Source Address Offset

TCD_SOFF3

TCD Signed Source Address Offset

TCD_SOFF4

TCD Signed Source Address Offset

TCD_SOFF5

TCD Signed Source Address Offset

TCD_SOFF6

TCD Signed Source Address Offset

TCD_SOFF7

TCD Signed Source Address Offset

TCD_SOFF8

TCD Signed Source Address Offset

TCD_SOFF9

TCD Signed Source Address Offset

TCD_SOFF10

TCD Signed Source Address Offset

TCD_SOFF11

TCD Signed Source Address Offset

TCD_SOFF12

TCD Signed Source Address Offset

TCD_SOFF13

TCD Signed Source Address Offset

TCD_SOFF14

TCD Signed Source Address Offset

TCD_SOFF15

TCD Signed Source Address Offset

TCD_SOFF16

TCD Signed Source Address Offset

TCD_SOFF17

TCD Signed Source Address Offset

TCD_SOFF18

TCD Signed Source Address Offset

TCD_SOFF19

TCD Signed Source Address Offset

TCD_SOFF20

TCD Signed Source Address Offset

TCD_SOFF21

TCD Signed Source Address Offset

TCD_SOFF22

TCD Signed Source Address Offset

TCD_SOFF23

TCD Signed Source Address Offset

TCD_SOFF24

TCD Signed Source Address Offset

TCD_SOFF25

TCD Signed Source Address Offset

TCD_SOFF26

TCD Signed Source Address Offset

TCD_SOFF27

TCD Signed Source Address Offset

TCD_SOFF28

TCD Signed Source Address Offset

TCD_SOFF29

TCD Signed Source Address Offset

TCD_SOFF30

TCD Signed Source Address Offset

TCD_SOFF31

TCD Signed Source Address Offset

Structs

Instance

RegisterBlock

ResetValues