Module imxrt_ral::imxrt105::peripherals::pmu::REG_3P0

Regulator 3P0 Register

Modules

BO_OFFSET
Control bits to adjust the regulator brownout offset voltage in 25mV steps

BO_VDD3P0
Status bit that signals when a brownout is detected on the regulator output.

ENABLE_BO
Control bit to enable the brownout circuitry in the regulator.

ENABLE_ILIMIT
Control bit to enable the current-limit circuitry in the regulator.

ENABLE_LINREG
Control bit to enable the regulator output to be set by the programmed target voltage setting and internal bandgap reference

OK_VDD3P0
Status bit that signals when the regulator output is ok. 1 = regulator output > brownout target

OUTPUT_TRG
Control bits to adjust the regulator output voltage

VBUS_SEL
Select input voltage source for LDO_3P0 from either USB_OTG1_VBUS or USB_OTG2_VBUS