Module imxrt_ral::imxrt105::peripherals::pwm::SMCTRL20::FORCE_SEL::RW

Read-write values

Constants

FORCE_SEL_0

0b000: The local force signal, CTRL2[FORCE], from this submodule is used to force updates.

FORCE_SEL_1

0b001: The master force signal from submodule 0 is used to force updates. This setting should not be used in submodule 0 as it will hold the FORCE OUTPUT signal to logic 0.

FORCE_SEL_2

0b010: The local reload signal from this submodule is used to force updates without regard to the state of LDOK.

FORCE_SEL_3

0b011: The master reload signal from submodule0 is used to force updates if LDOK is set. This setting should not be used in submodule0 as it will hold the FORCE OUTPUT signal to logic 0.

FORCE_SEL_4

0b100: The local sync signal from this submodule is used to force updates.

FORCE_SEL_5

0b101: The master sync signal from submodule0 is used to force updates. This setting should not be used in submodule0 as it will hold the FORCE OUTPUT signal to logic 0.

FORCE_SEL_6

0b110: The external force signal, EXT_FORCE, from outside the PWM module causes updates.

FORCE_SEL_7

0b111: The external sync signal, EXT_SYNC, from outside the PWM module causes updates.