Module imxrt_ral::imxrt105::peripherals::snvs::HPSVSR

SNVS_HP Security Violation Status Register

Modules

LP_SEC_VIO
LP Security Violation A security volation was detected in the SNVS low power section.

SV0
Security Violation 0 security violation was detected.

SV1
Security Violation 1 security violation was detected.

SV2
Security Violation 2 security violation was detected.

SV3
Security Violation 3 security violation was detected.

SV4
Security Violation 4 security violation was detected.

SV5
Security Violation 5 security violation was detected.

SW_FSV
Software Fatal Security Violation This bit is a read-only copy of the SW_FSV bit in the HP Command Register

SW_LPSV
LP Software Security Violation This bit is a read-only copy of the SW_LPSV bit in the HP Command Register

SW_SV
Software Security Violation This bit is a read-only copy of the SW_SV bit in the HP Command Register

ZMK_ECC_FAIL
Zeroizable Master Key Error Correcting Code Check Failure When set, this bit triggers a bad key violation to the SSM and a security violation to the SNVS_LP section, which clears security sensitive data

ZMK_SYNDROME
Zeroizable Master Key Syndrome The ZMK syndrome indicates the single-bit error location and parity for the ZMK register