Module imxrt_ral::imxrt105::peripherals::snvs::LPTDCR

SNVS_LP Tamper Detectors Configuration Register

Modules

ET1P
External Tampering 1 Polarity This bit is used to determine the polarity of external tamper 1.

ET1_EN
External Tampering 1 Enable When set, external tampering 1 detection generates an LP security violation

MCR_EN
MC Rollover Enable When set, an MC Rollover event generates an LP security violation.

OSCB
Oscillator Bypass When OSCB=1 the osc_bypass signal is asserted

PFD_OBSERV
System Power Fail Detector (PFD) Observability Flop The asynchronous reset input of this flop is connected directly to the inverted output of the PFD analog circuitry (external to the SNVS block)

POR_OBSERV
Power On Reset (POR) Observability Flop The asynchronous reset input of this flop is connected directly to the output of the POR analog circuitry (external to the SNVS

SRTCR_EN
SRTC Rollover Enable When set, an SRTC rollover event generates an LP security violation.