Module imxrt_ral::imxrt105::peripherals::usb::GPTIMER0CTRL

General Purpose Timer #0 Controller

Modules

GPTCNT
General Purpose Timer Counter. This field is the count value of the countdown timer.

GPTMODE
General Purpose Timer Mode In one shot mode, the timer will count down to zero, generate an interrupt, and stop until the counter is reset by software; In repeat mode, the timer will count down to zero, generate an interrupt and automatically reload the counter value from GPTLD bits to start again

GPTRST
General Purpose Timer Reset

GPTRUN
General Purpose Timer Run GPTCNT bits are not effected when setting or clearing this bit.