Module imxrt_ral::imxrt105::peripherals::usb::PORTSC1

Port Status & Control

Modules

CCS
Current Connect Status-Read Only

CSC
Connect Status Change-R/WC

FPR
Force Port Resume -Read/Write

HSP
High-Speed Port - Read Only

LS
Line Status-Read Only

OCA
Over-current Active-Read Only

OCC
Over-current Change-R/WC

PE
Port Enabled/Disabled-Read/Write

PEC
Port Enable/Disable Change-R/WC

PFSC
Port Force Full Speed Connect - Read/Write

PHCD
PHY Low Power Suspend - Clock Disable (PLPSCD) - Read/Write

PIC
Port Indicator Control - Read/Write

PO
Port Owner-Read/Write

PP
Port Power (PP)-Read/Write or Read Only

PR
Port Reset - Read/Write or Read Only

PSPD
Port Speed - Read Only. This register field indicates the speed at which the port is operating.

PTC
Port Test Control - Read/Write

PTS_1
All USB port interface modes are listed in this field description, but not all are supported

PTS_2
See description at bits 31-30

PTW
Parallel Transceiver Width This bit has no effect if serial interface engine is used

STS
Serial Transceiver Select 1 Serial Interface Engine is selected 0 Parallel Interface signals is selected Serial Interface Engine can be used in combination with UTMI+/ULPI physical interface to provide FS/LS signaling instead of the parallel interface signals

SUSP
Suspend - Read/Write or Read Only

WKCN
Wake on Connect Enable (WKCNNT_E) - Read/Write

WKDC
Wake on Disconnect Enable (WKDSCNNT_E) - Read/Write

WKOC
Wake on Over-current Enable (WKOC_E) - Read/Write