Struct imxrt_ral::imxrt106::imxrt1064::dma0::DMA0

pub struct DMA0 { /* private fields */ }

The DMA0 peripheral instance.

This is a new type only for documentation purposes. When compiling for a target, this is defined as

pub type DMA0 = Instance<0>;

Implementations

impl DMA0

Access functions for the DMA0 peripheral instance

pub const reset: ResetValues = ResetValues { CR: 0x00000400, ES: 0x00000000, ERQ: 0x00000000, EEI: 0x00000000, CEEI: 0x00000000, SEEI: 0x00000000, CERQ: 0x00000000, SERQ: 0x00000000, CDNE: 0x00000000, SSRT: 0x00000000, CERR: 0x00000000, CINT: 0x00000000, INT: 0x00000000, ERR: 0x00000000, HRS: 0x00000000, EARS: 0x00000000, DCHPRI3: 0x00000003, DCHPRI2: 0x00000002, DCHPRI1: 0x00000001, DCHPRI0: 0x00000000, DCHPRI7: 0x00000007, DCHPRI6: 0x00000006, DCHPRI5: 0x00000005, DCHPRI4: 0x00000004, DCHPRI11: 0x0000000B, DCHPRI10: 0x0000000A, DCHPRI9: 0x00000009, DCHPRI8: 0x00000008, DCHPRI15: 0x0000000F, DCHPRI14: 0x0000000E, DCHPRI13: 0x0000000D, DCHPRI12: 0x0000000C, DCHPRI19: 0x00000013, DCHPRI18: 0x00000012, DCHPRI17: 0x00000011, DCHPRI16: 0x00000010, DCHPRI23: 0x00000017, DCHPRI22: 0x00000016, DCHPRI21: 0x00000015, DCHPRI20: 0x00000014, DCHPRI27: 0x0000001B, DCHPRI26: 0x0000001A, DCHPRI25: 0x00000019, DCHPRI24: 0x00000018, DCHPRI31: 0x0000001F, DCHPRI30: 0x0000001E, DCHPRI29: 0x0000001D, DCHPRI28: 0x0000001C, TCD_SADDR0: 0x00000000, TCD_SOFF0: 0x00000000, TCD_ATTR0: 0x00000000, TCD_NBYTES_ML: 0x00000000, TCD_SLAST0: 0x00000000, TCD_DADDR0: 0x00000000, TCD_DOFF0: 0x00000000, TCD_CITER_ELINK: 0x00000000, TCD_DLASTSGA0: 0x00000000, TCD_CSR0: 0x00000000, TCD_BITER_ELINK: 0x00000000, TCD_SADDR1: 0x00000000, TCD_SOFF1: 0x00000000, TCD_ATTR1: 0x00000000, TCD_NBYTES_MLNO1: 0x00000000, TCD_SLAST1: 0x00000000, TCD_DADDR1: 0x00000000, TCD_DOFF1: 0x00000000, TCD_CITER_ELINKNO1: 0x00000000, TCD_DLASTSGA1: 0x00000000, TCD_CSR1: 0x00000000, TCD_BITER_ELINKNO1: 0x00000000, TCD_SADDR2: 0x00000000, TCD_SOFF2: 0x00000000, TCD_ATTR2: 0x00000000, TCD_NBYTES_MLNO2: 0x00000000, TCD_SLAST2: 0x00000000, TCD_DADDR2: 0x00000000, TCD_DOFF2: 0x00000000, TCD_CITER_ELINKNO2: 0x00000000, TCD_DLASTSGA2: 0x00000000, TCD_CSR2: 0x00000000, TCD_BITER_ELINKNO2: 0x00000000, TCD_SADDR3: 0x00000000, TCD_SOFF3: 0x00000000, TCD_ATTR3: 0x00000000, TCD_NBYTES_MLNO3: 0x00000000, TCD_SLAST3: 0x00000000, TCD_DADDR3: 0x00000000, TCD_DOFF3: 0x00000000, TCD_CITER_ELINKNO3: 0x00000000, TCD_DLASTSGA3: 0x00000000, TCD_CSR3: 0x00000000, TCD_BITER_ELINKNO3: 0x00000000, TCD_SADDR4: 0x00000000, TCD_SOFF4: 0x00000000, TCD_ATTR4: 0x00000000, TCD_NBYTES_MLNO4: 0x00000000, TCD_SLAST4: 0x00000000, TCD_DADDR4: 0x00000000, TCD_DOFF4: 0x00000000, TCD_CITER_ELINKNO4: 0x00000000, TCD_DLASTSGA4: 0x00000000, TCD_CSR4: 0x00000000, TCD_BITER_ELINKNO4: 0x00000000, TCD_SADDR5: 0x00000000, TCD_SOFF5: 0x00000000, TCD_ATTR5: 0x00000000, TCD_NBYTES_MLNO5: 0x00000000, TCD_SLAST5: 0x00000000, TCD_DADDR5: 0x00000000, TCD_DOFF5: 0x00000000, TCD_CITER_ELINKNO5: 0x00000000, TCD_DLASTSGA5: 0x00000000, TCD_CSR5: 0x00000000, TCD_BITER_ELINKNO5: 0x00000000, TCD_SADDR6: 0x00000000, TCD_SOFF6: 0x00000000, TCD_ATTR6: 0x00000000, TCD_NBYTES_MLNO6: 0x00000000, TCD_SLAST6: 0x00000000, TCD_DADDR6: 0x00000000, TCD_DOFF6: 0x00000000, TCD_CITER_ELINKNO6: 0x00000000, TCD_DLASTSGA6: 0x00000000, TCD_CSR6: 0x00000000, TCD_BITER_ELINKNO6: 0x00000000, TCD_SADDR7: 0x00000000, TCD_SOFF7: 0x00000000, TCD_ATTR7: 0x00000000, TCD_NBYTES_MLNO7: 0x00000000, TCD_SLAST7: 0x00000000, TCD_DADDR7: 0x00000000, TCD_DOFF7: 0x00000000, TCD_CITER_ELINKNO7: 0x00000000, TCD_DLASTSGA7: 0x00000000, TCD_CSR7: 0x00000000, TCD_BITER_ELINKNO7: 0x00000000, TCD_SADDR8: 0x00000000, TCD_SOFF8: 0x00000000, TCD_ATTR8: 0x00000000, TCD_NBYTES_MLNO8: 0x00000000, TCD_SLAST8: 0x00000000, TCD_DADDR8: 0x00000000, TCD_DOFF8: 0x00000000, TCD_CITER_ELINKNO8: 0x00000000, TCD_DLASTSGA8: 0x00000000, TCD_CSR8: 0x00000000, TCD_BITER_ELINKNO8: 0x00000000, TCD_SADDR9: 0x00000000, TCD_SOFF9: 0x00000000, TCD_ATTR9: 0x00000000, TCD_NBYTES_MLNO9: 0x00000000, TCD_SLAST9: 0x00000000, TCD_DADDR9: 0x00000000, TCD_DOFF9: 0x00000000, TCD_CITER_ELINKNO9: 0x00000000, TCD_DLASTSGA9: 0x00000000, TCD_CSR9: 0x00000000, TCD_BITER_ELINKNO9: 0x00000000, TCD_SADDR10: 0x00000000, TCD_SOFF10: 0x00000000, TCD_ATTR10: 0x00000000, TCD_NBYTES_MLNO10: 0x00000000, TCD_SLAST10: 0x00000000, TCD_DADDR10: 0x00000000, TCD_DOFF10: 0x00000000, TCD_CITER_ELINKNO10: 0x00000000, TCD_DLASTSGA10: 0x00000000, TCD_CSR10: 0x00000000, TCD_BITER_ELINKNO10: 0x00000000, TCD_SADDR11: 0x00000000, TCD_SOFF11: 0x00000000, TCD_ATTR11: 0x00000000, TCD_NBYTES_MLNO11: 0x00000000, TCD_SLAST11: 0x00000000, TCD_DADDR11: 0x00000000, TCD_DOFF11: 0x00000000, TCD_CITER_ELINKNO11: 0x00000000, TCD_DLASTSGA11: 0x00000000, TCD_CSR11: 0x00000000, TCD_BITER_ELINKNO11: 0x00000000, TCD_SADDR12: 0x00000000, TCD_SOFF12: 0x00000000, TCD_ATTR12: 0x00000000, TCD_NBYTES_MLNO12: 0x00000000, TCD_SLAST12: 0x00000000, TCD_DADDR12: 0x00000000, TCD_DOFF12: 0x00000000, TCD_CITER_ELINKNO12: 0x00000000, TCD_DLASTSGA12: 0x00000000, TCD_CSR12: 0x00000000, TCD_BITER_ELINKNO12: 0x00000000, TCD_SADDR13: 0x00000000, TCD_SOFF13: 0x00000000, TCD_ATTR13: 0x00000000, TCD_NBYTES_MLNO13: 0x00000000, TCD_SLAST13: 0x00000000, TCD_DADDR13: 0x00000000, TCD_DOFF13: 0x00000000, TCD_CITER_ELINKNO13: 0x00000000, TCD_DLASTSGA13: 0x00000000, TCD_CSR13: 0x00000000, TCD_BITER_ELINKNO13: 0x00000000, TCD_SADDR14: 0x00000000, TCD_SOFF14: 0x00000000, TCD_ATTR14: 0x00000000, TCD_NBYTES_MLNO14: 0x00000000, TCD_SLAST14: 0x00000000, TCD_DADDR14: 0x00000000, TCD_DOFF14: 0x00000000, TCD_CITER_ELINKNO14: 0x00000000, TCD_DLASTSGA14: 0x00000000, TCD_CSR14: 0x00000000, TCD_BITER_ELINKNO14: 0x00000000, TCD_SADDR15: 0x00000000, TCD_SOFF15: 0x00000000, TCD_ATTR15: 0x00000000, TCD_NBYTES_MLNO15: 0x00000000, TCD_SLAST15: 0x00000000, TCD_DADDR15: 0x00000000, TCD_DOFF15: 0x00000000, TCD_CITER_ELINKNO15: 0x00000000, TCD_DLASTSGA15: 0x00000000, TCD_CSR15: 0x00000000, TCD_BITER_ELINKNO15: 0x00000000, TCD_SADDR16: 0x00000000, TCD_SOFF16: 0x00000000, TCD_ATTR16: 0x00000000, TCD_NBYTES_MLNO16: 0x00000000, TCD_SLAST16: 0x00000000, TCD_DADDR16: 0x00000000, TCD_DOFF16: 0x00000000, TCD_CITER_ELINKNO16: 0x00000000, TCD_DLASTSGA16: 0x00000000, TCD_CSR16: 0x00000000, TCD_BITER_ELINKNO16: 0x00000000, TCD_SADDR17: 0x00000000, TCD_SOFF17: 0x00000000, TCD_ATTR17: 0x00000000, TCD_NBYTES_MLNO17: 0x00000000, TCD_SLAST17: 0x00000000, TCD_DADDR17: 0x00000000, TCD_DOFF17: 0x00000000, TCD_CITER_ELINKNO17: 0x00000000, TCD_DLASTSGA17: 0x00000000, TCD_CSR17: 0x00000000, TCD_BITER_ELINKNO17: 0x00000000, TCD_SADDR18: 0x00000000, TCD_SOFF18: 0x00000000, TCD_ATTR18: 0x00000000, TCD_NBYTES_MLNO18: 0x00000000, TCD_SLAST18: 0x00000000, TCD_DADDR18: 0x00000000, TCD_DOFF18: 0x00000000, TCD_CITER_ELINKNO18: 0x00000000, TCD_DLASTSGA18: 0x00000000, TCD_CSR18: 0x00000000, TCD_BITER_ELINKNO18: 0x00000000, TCD_SADDR19: 0x00000000, TCD_SOFF19: 0x00000000, TCD_ATTR19: 0x00000000, TCD_NBYTES_MLNO19: 0x00000000, TCD_SLAST19: 0x00000000, TCD_DADDR19: 0x00000000, TCD_DOFF19: 0x00000000, TCD_CITER_ELINKNO19: 0x00000000, TCD_DLASTSGA19: 0x00000000, TCD_CSR19: 0x00000000, TCD_BITER_ELINKNO19: 0x00000000, TCD_SADDR20: 0x00000000, TCD_SOFF20: 0x00000000, TCD_ATTR20: 0x00000000, TCD_NBYTES_MLNO20: 0x00000000, TCD_SLAST20: 0x00000000, TCD_DADDR20: 0x00000000, TCD_DOFF20: 0x00000000, TCD_CITER_ELINKNO20: 0x00000000, TCD_DLASTSGA20: 0x00000000, TCD_CSR20: 0x00000000, TCD_BITER_ELINKNO20: 0x00000000, TCD_SADDR21: 0x00000000, TCD_SOFF21: 0x00000000, TCD_ATTR21: 0x00000000, TCD_NBYTES_MLNO21: 0x00000000, TCD_SLAST21: 0x00000000, TCD_DADDR21: 0x00000000, TCD_DOFF21: 0x00000000, TCD_CITER_ELINKNO21: 0x00000000, TCD_DLASTSGA21: 0x00000000, TCD_CSR21: 0x00000000, TCD_BITER_ELINKNO21: 0x00000000, TCD_SADDR22: 0x00000000, TCD_SOFF22: 0x00000000, TCD_ATTR22: 0x00000000, TCD_NBYTES_MLNO22: 0x00000000, TCD_SLAST22: 0x00000000, TCD_DADDR22: 0x00000000, TCD_DOFF22: 0x00000000, TCD_CITER_ELINKNO22: 0x00000000, TCD_DLASTSGA22: 0x00000000, TCD_CSR22: 0x00000000, TCD_BITER_ELINKNO22: 0x00000000, TCD_SADDR23: 0x00000000, TCD_SOFF23: 0x00000000, TCD_ATTR23: 0x00000000, TCD_NBYTES_MLNO23: 0x00000000, TCD_SLAST23: 0x00000000, TCD_DADDR23: 0x00000000, TCD_DOFF23: 0x00000000, TCD_CITER_ELINKNO23: 0x00000000, TCD_DLASTSGA23: 0x00000000, TCD_CSR23: 0x00000000, TCD_BITER_ELINKNO23: 0x00000000, TCD_SADDR24: 0x00000000, TCD_SOFF24: 0x00000000, TCD_ATTR24: 0x00000000, TCD_NBYTES_MLNO24: 0x00000000, TCD_SLAST24: 0x00000000, TCD_DADDR24: 0x00000000, TCD_DOFF24: 0x00000000, TCD_CITER_ELINKNO24: 0x00000000, TCD_DLASTSGA24: 0x00000000, TCD_CSR24: 0x00000000, TCD_BITER_ELINKNO24: 0x00000000, TCD_SADDR25: 0x00000000, TCD_SOFF25: 0x00000000, TCD_ATTR25: 0x00000000, TCD_NBYTES_MLNO25: 0x00000000, TCD_SLAST25: 0x00000000, TCD_DADDR25: 0x00000000, TCD_DOFF25: 0x00000000, TCD_CITER_ELINKNO25: 0x00000000, TCD_DLASTSGA25: 0x00000000, TCD_CSR25: 0x00000000, TCD_BITER_ELINKNO25: 0x00000000, TCD_SADDR26: 0x00000000, TCD_SOFF26: 0x00000000, TCD_ATTR26: 0x00000000, TCD_NBYTES_MLNO26: 0x00000000, TCD_SLAST26: 0x00000000, TCD_DADDR26: 0x00000000, TCD_DOFF26: 0x00000000, TCD_CITER_ELINKNO26: 0x00000000, TCD_DLASTSGA26: 0x00000000, TCD_CSR26: 0x00000000, TCD_BITER_ELINKNO26: 0x00000000, TCD_SADDR27: 0x00000000, TCD_SOFF27: 0x00000000, TCD_ATTR27: 0x00000000, TCD_NBYTES_MLNO27: 0x00000000, TCD_SLAST27: 0x00000000, TCD_DADDR27: 0x00000000, TCD_DOFF27: 0x00000000, TCD_CITER_ELINKNO27: 0x00000000, TCD_DLASTSGA27: 0x00000000, TCD_CSR27: 0x00000000, TCD_BITER_ELINKNO27: 0x00000000, TCD_SADDR28: 0x00000000, TCD_SOFF28: 0x00000000, TCD_ATTR28: 0x00000000, TCD_NBYTES_MLNO28: 0x00000000, TCD_SLAST28: 0x00000000, TCD_DADDR28: 0x00000000, TCD_DOFF28: 0x00000000, TCD_CITER_ELINKNO28: 0x00000000, TCD_DLASTSGA28: 0x00000000, TCD_CSR28: 0x00000000, TCD_BITER_ELINKNO28: 0x00000000, TCD_SADDR29: 0x00000000, TCD_SOFF29: 0x00000000, TCD_ATTR29: 0x00000000, TCD_NBYTES_MLNO29: 0x00000000, TCD_SLAST29: 0x00000000, TCD_DADDR29: 0x00000000, TCD_DOFF29: 0x00000000, TCD_CITER_ELINKNO29: 0x00000000, TCD_DLASTSGA29: 0x00000000, TCD_CSR29: 0x00000000, TCD_BITER_ELINKNO29: 0x00000000, TCD_SADDR30: 0x00000000, TCD_SOFF30: 0x00000000, TCD_ATTR30: 0x00000000, TCD_NBYTES_MLNO30: 0x00000000, TCD_SLAST30: 0x00000000, TCD_DADDR30: 0x00000000, TCD_DOFF30: 0x00000000, TCD_CITER_ELINKNO30: 0x00000000, TCD_DLASTSGA30: 0x00000000, TCD_CSR30: 0x00000000, TCD_BITER_ELINKNO30: 0x00000000, TCD_SADDR31: 0x00000000, TCD_SOFF31: 0x00000000, TCD_ATTR31: 0x00000000, TCD_NBYTES_MLNO31: 0x00000000, TCD_SLAST31: 0x00000000, TCD_DADDR31: 0x00000000, TCD_DOFF31: 0x00000000, TCD_CITER_ELINKNO31: 0x00000000, TCD_DLASTSGA31: 0x00000000, TCD_CSR31: 0x00000000, TCD_BITER_ELINKNO31: 0x00000000, }

Reset values for each field in DMA0

pub fn take() -> Option<Self>

Safe access to DMA0

This function returns Some(Instance) if this instance is not currently taken, and None if it is. This ensures that if you do get Some(Instance), you are ensured unique access to the peripheral and there cannot be data races (unless other code uses unsafe, of course). You can then pass the Instance around to other functions as required. When you’re done with it, you can call release(instance) to return it.

Instance itself dereferences to a RegisterBlock, which provides access to the peripheral’s registers.

pub fn release(_: Self)

Release exclusive access to DMA0

This function allows you to return an Instance so that it is available to take() again. This function will panic if you return a different Instance or if this instance is not already taken.

pub unsafe fn steal() -> Self

Unsafely steal DMA0

This function is similar to take() but forcibly takes the Instance, marking it as taken irregardless of its previous state.

impl DMA0

pub const INTERRUPTS: [Interrupt; 0] = []

The interrupts associated with DMA0

Note: the values are invalid for a documentation build.