Module imxrt_ral::imxrt106::peripherals::can::IFLAG1

Interrupt Flags 1 Register

Modules

BUF4TO0I

If the Rx FIFO is not enabled, these bits flag the interrupts for MB0 to MB4

BUF5I

If the Rx FIFO is not enabled, this bit flags the interrupt for MB5

BUF6I

If the Rx FIFO is not enabled, this bit flags the interrupt for MB6

BUF7I

If the Rx FIFO is not enabled, this bit flags the interrupt for MB7

BUF31TO8I

Each bit flags the respective FLEXCAN Message Buffer (MB8 to MB31) interrupt.