Module imxrt_ral::imxrt106::peripherals::can3::CS42

CS42 and MB9_64B_WORD4 CS42: CS42 and MB42_8B_CS CS42: CS42 and MB28_16B_CS CS42: CS42 and MB16_32B_WORD6 CS42: Message Buffer 42 CS Register MB16_32B_WORD6: Message Buffer 16 WORD_32B Register MB28_16B_CS: Message Buffer 28 CS Register MB42_8B_CS: Message Buffer 42 CS Register MB9_64B_WORD4: Message Buffer 9 WORD_64B Register

Modules

BRS

Bit Rate Switch. This bit defines whether the bit rate is switched inside a CAN FD format frame.

CODE

Message Buffer Code. This 4-bit field can be accessed (read or write) by the CPU and by the FlexCAN module itself, as part of the message buffer matching and arbitration process.

DATA_BYTE_16

Data byte 3 of Rx/Tx frame.

DATA_BYTE_17

Data byte 2 of Rx/Tx frame.

DATA_BYTE_18

Data byte 1 of Rx/Tx frame.

DATA_BYTE_19

Data byte 0 of Rx/Tx frame.

DATA_BYTE_24

Data byte 3 of Rx/Tx frame.

DATA_BYTE_25

Data byte 2 of Rx/Tx frame.

DATA_BYTE_26

Data byte 1 of Rx/Tx frame.

DATA_BYTE_27

Data byte 0 of Rx/Tx frame.

DLC

Length of the data to be stored/transmitted.

EDL

Extended Data Length. This bit distinguishes between CAN format and CAN FD format frames. The EDL bit must not be set for Message Buffers configured to RANSWER with code field 0b1010.

ESI

Error State Indicator. This bit indicates if the transmitting node is error active or error passive.

IDE

ID Extended. One/zero for extended/standard format frame.

RTR

Remote Transmission Request. One/zero for remote/data frame.

SRR

Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP

Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.