Module imxrt_ral::imxrt106::peripherals::can3::CTRL2

Control 2 register

Modules

BOFFDONEMSK
Bus Off Done Interrupt Mask

BTE
Bit Timing Expansion enable

EACEN
Entire Frame Arbitration Field Comparison Enable For Rx Mailboxes

EDFLTDIS
Edge Filter Disable

ERRMSK_FAST
Error Interrupt Mask for errors detected in the Data Phase of fast CAN FD frames

ISOCANFDEN
ISO CAN FD Enable

MBTSBASE
Message Buffer Time Stamp Base

MRP
Mailboxes Reception Priority

PREXCEN
Protocol Exception Enable

RFFN
Number Of Legacy Rx FIFO Filters

RRS
Remote Request Storing

TASD
Tx Arbitration Start Delay

TIMER_SRC
Timer Source

TSTAMPCAP
Time Stamp Capture Point