Module imxrt_ral::imxrt106::peripherals::can3::MB10_32B_WORD0

MB10_32B_WORD0 and WORD025 MB10_32B_WORD0: MB10_32B_WORD0 and MB5_64B_WORD10 MB10_32B_WORD0: MB10_32B_WORD0 and MB25_8B_WORD0 MB10_32B_WORD0: MB10_32B_WORD0 and MB17_16B_CS MB10_32B_WORD0: Message Buffer 10 WORD_32B Register MB17_16B_CS: Message Buffer 17 CS Register MB25_8B_WORD0: Message Buffer 25 WORD_8B Register MB5_64B_WORD10: Message Buffer 5 WORD_64B Register WORD025: Message Buffer 25 WORD0 Register

Modules

BRS

Bit Rate Switch. This bit defines whether the bit rate is switched inside a CAN FD format frame.

CODE

Message Buffer Code. This 4-bit field can be accessed (read or write) by the CPU and by the FlexCAN module itself, as part of the message buffer matching and arbitration process.

DATA_BYTE_0

Data byte 3 of Rx/Tx frame.

DATA_BYTE_1

Data byte 2 of Rx/Tx frame.

DATA_BYTE_2

Data byte 1 of Rx/Tx frame.

DATA_BYTE_3

Data byte 0 of Rx/Tx frame.

DATA_BYTE_40

Data byte 3 of Rx/Tx frame.

DATA_BYTE_41

Data byte 2 of Rx/Tx frame.

DATA_BYTE_42

Data byte 1 of Rx/Tx frame.

DATA_BYTE_43

Data byte 0 of Rx/Tx frame.

DLC

Length of the data to be stored/transmitted.

EDL

Extended Data Length. This bit distinguishes between CAN format and CAN FD format frames. The EDL bit must not be set for Message Buffers configured to RANSWER with code field 0b1010.

ESI

Error State Indicator. This bit indicates if the transmitting node is error active or error passive.

IDE

ID Extended. One/zero for extended/standard format frame.

RTR

Remote Transmission Request. One/zero for remote/data frame.

SRR

Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP

Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.