Module imxrt_ral::imxrt106::peripherals::can3::MB12_64B_WORD12

MB12_64B_WORD12 and WORD057 MB12_64B_WORD12: MB12_64B_WORD12 and MB57_8B_WORD0 MB12_64B_WORD12: MB12_64B_WORD12 and MB38_16B_WORD0 MB12_64B_WORD12: MB12_64B_WORD12 and MB23_32B_CS MB12_64B_WORD12: Message Buffer 12 WORD_64B Register MB23_32B_CS: Message Buffer 23 CS Register MB38_16B_WORD0: Message Buffer 38 WORD_16B Register MB57_8B_WORD0: Message Buffer 57 WORD_8B Register WORD057: Message Buffer 57 WORD0 Register

Modules

BRS
Bit Rate Switch. This bit defines whether the bit rate is switched inside a CAN FD format frame.

CODE
Message Buffer Code. This 4-bit field can be accessed (read or write) by the CPU and by the FlexCAN module itself, as part of the message buffer matching and arbitration process.

DATA_BYTE_0
Data byte 3 of Rx/Tx frame.

DATA_BYTE_1
Data byte 2 of Rx/Tx frame.

DATA_BYTE_2
Data byte 1 of Rx/Tx frame.

DATA_BYTE_3
Data byte 0 of Rx/Tx frame.

DATA_BYTE_48
Data byte 3 of Rx/Tx frame.

DATA_BYTE_49
Data byte 2 of Rx/Tx frame.

DATA_BYTE_50
Data byte 1 of Rx/Tx frame.

DATA_BYTE_51
Data byte 0 of Rx/Tx frame.

DLC
Length of the data to be stored/transmitted.

EDL
Extended Data Length. This bit distinguishes between CAN format and CAN FD format frames. The EDL bit must not be set for Message Buffers configured to RANSWER with code field 0b1010.

ESI
Error State Indicator. This bit indicates if the transmitting node is error active or error passive.

IDE
ID Extended. One/zero for extended/standard format frame.

RTR
Remote Transmission Request. One/zero for remote/data frame.

SRR
Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP
Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.