Module imxrt_ral::imxrt106::peripherals::can3::MB13_8B_WORD0

MB13_8B_WORD0 and WORD013 MB13_8B_WORD0: MB13_8B_WORD0 and MB9_16B_CS MB13_8B_WORD0: MB13_8B_WORD0 and MB5_32B_WORD2 MB13_8B_WORD0: MB13_8B_WORD0 and MB3_64B_CS MB13_8B_WORD0: Message Buffer 13 WORD_8B Register MB3_64B_CS: Message Buffer 3 CS Register MB5_32B_WORD2: Message Buffer 5 WORD_32B Register MB9_16B_CS: Message Buffer 9 CS Register WORD013: Message Buffer 13 WORD0 Register

Modules

BRS

Bit Rate Switch. This bit defines whether the bit rate is switched inside a CAN FD format frame.

CODE

Message Buffer Code. This 4-bit field can be accessed (read or write) by the CPU and by the FlexCAN module itself, as part of the message buffer matching and arbitration process.

DATA_BYTE_0

Data byte 3 of Rx/Tx frame.

DATA_BYTE_1

Data byte 2 of Rx/Tx frame.

DATA_BYTE_2

Data byte 1 of Rx/Tx frame.

DATA_BYTE_3

Data byte 0 of Rx/Tx frame.

DATA_BYTE_8

Data byte 3 of Rx/Tx frame.

DATA_BYTE_9

Data byte 2 of Rx/Tx frame.

DATA_BYTE_10

Data byte 1 of Rx/Tx frame.

DATA_BYTE_11

Data byte 0 of Rx/Tx frame.

DLC

Length of the data to be stored/transmitted.

EDL

Extended Data Length. This bit distinguishes between CAN format and CAN FD format frames. The EDL bit must not be set for Message Buffers configured to RANSWER with code field 0b1010.

ESI

Error State Indicator. This bit indicates if the transmitting node is error active or error passive.

IDE

ID Extended. One/zero for extended/standard format frame.

RTR

Remote Transmission Request. One/zero for remote/data frame.

SRR

Substitute Remote Request. Contains a fixed recessive bit.

TIME_STAMP

Free-Running Counter Time stamp. This 16-bit field is a copy of the Free-Running Timer, captured for Tx and Rx frames at the time when the beginning of the Identifier field appears on the CAN bus.