Module imxrt_ral::imxrt106::peripherals::pxp::CTRL

Control Register 0

Modules

BLOCK_SIZE
Select the block size to process.

CLKGATE
This bit must be set to zero for normal operation

ENABLE
Enables PXP operation with specified parameters

ENABLE_LCD_HANDSHAKE
Enable handshake with LCD controller

EN_REPEAT
Enable the PXP to run continuously

HFLIP
Indicates that the output buffer should be flipped horizontally (effect applied before rotation).

IRQ_ENABLE
Interrupt enable

NEXT_IRQ_ENABLE
Next command interrupt enable

ROTATE
Indicates the clockwise rotation to be applied at the output buffer

ROT_POS
This bit controls where rotation will occur in the PXP datapath

RSVD0
Reserved, always set to zero.

RSVD1
Reserved, always set to zero.

RSVD3
Reserved, always set to zero.

RSVD4
Reserved, always set to zero.

SFTRST
Set this bit to zero to enable normal PXP operation

VFLIP
Indicates that the output buffer should be flipped vertically (effect applied before rotation).