Module imxrt_ral::imxrt106::peripherals::pxp::OUT_CTRL

Output Buffer Control Register

Modules

ALPHA
When generating an output buffer with an alpha component, the value in this field will be used when enabled to override the alpha passed through the pixel data pipeline

ALPHA_OUTPUT
Indicates that alpha component in output buffer pixels should be overwritten by PXP_OUT_CTRL[ALPHA]

FORMAT
Output framebuffer format

INTERLACED_OUTPUT
Determines how the PXP writes it's output data

RSVD0
Reserved, always set to zero.

RSVD1
Reserved, always set to zero.