Module imxrt_ral::imxrt106::peripherals::pxp::STAT

Status Register

Modules

AXI_ERROR_ID

Indicates the AXI ID of the failing bus operation.

AXI_READ_ERROR

Indicates PXP encountered an AXI read error and processing has been terminated.

AXI_WRITE_ERROR

Indicates PXP encountered an AXI write error and processing has been terminated.

BLOCKX

Indicates the X coordinate of the block currently being rendered.

BLOCKY

Indicates the X coordinate of the block currently being rendered.

IRQ

Indicates current PXP interrupt status

LUT_DMA_LOAD_DONE_IRQ

Indicates that the LUT DMA transfer has completed.

NEXT_IRQ

Indicates that a command issued with the “Next Command” functionality has been issued and that a new command may be initiated with a write to the PXP_NEXT register

RSVD2

Reserved, always set to zero.