Module imxrt_ral::imxrt106::peripherals::usb::HCSPARAMS

Host Controller Structural Parameters

Modules

N_CC
Number of Companion Controller (N_CC)

N_PCC
Number of Ports per Companion Controller This field indicates the number of ports supported per internal Companion Controller

N_PORTS
Number of downstream ports

N_PTT
Number of Ports per Transaction Translator (N_PTT)

N_TT
Number of Transaction Translators (N_TT)

PI
Port Indicators (P INDICATOR) This bit indicates whether the ports support port indicator control

PPC
Port Power Control This field indicates whether the host controller implementation includes port power control