Macro imxrt_ral::write_reg

macro_rules! write_reg {
    ( $periph:path, $instance:expr, $reg:ident, $( $field:ident : $value:expr ),+ ) => { ... };
    ( $periph:path, $instance:expr, $reg:ident, $value:expr ) => { ... };
}

Write to a RWRegister or UnsafeRWRegister.

Examples

// Safely acquire the peripheral instance (will panic if already acquired)
let gpio = imxrt_ral::gpio::GPIO1::take().unwrap();

// Write some value to the register.
write_reg!(imxrt_ral::gpio, gpio, ICR1, 0b11<<6);

// Write values to specific fields. Unspecified fields are written to 0.
write_reg!(imxrt_ral::gpio, gpio, ICR1, ICR3: FALLING_EDGE, ICR4: LOW_LEVEL);

// Unsafe access without requiring you to first `take()` the instance
unsafe { write_reg!(imxrt_ral::gpio, GPIO1, ICR1, ICR3: FALLING_EDGE, ICR4: LOW_LEVEL) };

Usage

Like modify_reg!, this macro can be used in two ways, either with a single value to write to the whole register, or with multiple fields each with their own value.

In both cases, the first arguments are:

the path to the peripheral module: imxrt_ral::gpio,

a reference to the instance of that peripheral: ‘gpio’ (anything which dereferences to RegisterBlock, such as Instance, &Instance, &RegisterBlock, or *const RegisterBlock),

the register you wish you access: ICR1 (a field on the RegisterBlock).

In the single-value usage, the final argument is just the value to write:

// Turn on the fourth pin (and turn everything else off).
write_reg!(imxrt_ral::gpio, gpio, DR, 1<<3);

Otherwise, the remaining arguments are each Field: Value pairs:

// Set ICR3 and ICR4, and everything else to 0.
write_reg!(imxrt_ral::gpio, gpio, ICR1, ICR3: 0b01, ICR4: 0b11);

For fields with annotated values, you can also specify a named value:

// As above, but with named values.
write_reg!(imxrt_ral::gpio, gpio, ICR1, ICR3: FALLING_EDGE, ICR4: LOW_LEVEL);

This macro expands to calling (*$instance).$register.write(value), where in the second usage, the value is computed as the bitwise OR of each field value, which are masked and shifted appropriately for the given field. The named values are brought into scope by use $peripheral::$register::$field::* for each field. The same constants could just be specified manually:

// As above, but being explicit about named values.
write_reg!(imxrt_ral::gpio, gpio, ICR1, ICR3: imxrt_ral::gpio::ICR1::ICR3::RW::FALLING_EDGE,
                                         ICR4: imxrt_ral::gpio::ICR1::ICR4::RW::LOW_LEVEL);

The fully expanded form is equivalent to:

// As above, but expanded.
(*gpio).ICR1.write(
    ((imxrt_ral::gpio::ICR1::ICR3::RW::FALLING_EDGE << imxrt_ral::gpio::ICR1::ICR3::offset)
     & imxrt_ral::gpio::ICR1::ICR3::mask)
    |
    ((imxrt_ral::gpio::ICR1::ICR4::RW::LOW_LEVEL << imxrt_ral::gpio::ICR1::ICR4::offset)
     & imxrt_ral::gpio::ICR1::ICR4::mask)
);

Safety

This macro will require an unsafe function or block when used with an UnsafeRWRegister, but not if used with RWRegister.

When run in an unsafe context, peripheral instances are directly accessible without requiring having called take() beforehand:

unsafe { write_reg!(imxrt_ral::gpio, GPIO1, ICR1, ICR3: FALLING_EDGE, ICR4: LOW_LEVEL) };

This works because GPIO1 is a *const RegisterBlock in the imxrt_ral::gpio module; and the macro brings such constants into scope and then dereferences the provided reference.