Module imxrt_ral::pwm::sm::SMCTRL::COMPMODE::RW

Constants

COMPMODE_0

The VAL* registers and the PWM counter are compared using an “equal to” method. This means that PWM edges are only produced when the counter is equal to one of the VAL* register values. This implies that a PWMA output that is high at the end of a period will maintain this state until a match with VAL3 clears the output in the following period.

COMPMODE_1

The VAL* registers and the PWM counter are compared using an “equal to or greater than” method. This means that PWM edges are produced when the counter is equal to or greater than one of the VAL* register values. This implies that a PWMA output that is high at the end of a period could go low at the start of the next period if the starting counter value is greater than (but not necessarily equal to) the new VAL3 value.