imxrt_ral::pxp::AS_CTRL::ALPHA_CTRL
Module RW

Constants

EMBEDDED

Indicates that the AS pixel alpha value will be used to blend the AS with PS. The ALPHA field is ignored.

MULTIPLY

Indicates that the value in the ALPHA field should be used to scale all pixel alpha values. Each pixel alpha is multiplied by the value in the ALPHA field.

OVERRIDE

Indicates that the value in the ALPHA field should be used instead of the alpha values present in the input pixels.

ROPS

Enable ROPs. The ROP field indicates an operation to be performed on the alpha surface and PS pixels.