Struct imxrt_ral::Instance

#[repr(transparent)]
pub struct Instance<T, const N: u8> { /* private fields */ }

An owned peripheral of type T, instance N.

Fabricating an Instance is always unsafe. An owner of an Instance may assume that

the underlying pointer points to a static register block of type T.

the instance number N properly describes the peripheral instance.

they own all registers pointed at by T.

Owners use this guarantee to safely access the peripheral registers. However, nothing guarantees any of these except for your diligence.

Constructing an Instance is zero cost. Additionally, Instance is transparent and amenable to null-pointer optimizations.

See the package-level documentation for more information on fabricating instances.

Safety of new().

By calling new(), you claim

ptr points to static memory that can be described by a type T.

The instance number N correctly describes ptr.

You are becoming the sole owner of this instance.

Safety of instance()

The various instance() methods handle safety concerns 1 and 2 from new(). By their construction, each instance() implementation provides a pointer to valid peripheral memory, and associates the correct N with that pointer. Therefore, you’re only responsible for ensuring safety concern 3 from new().

Implementations

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 5>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 6>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 7>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 8>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 9>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 5>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 6>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 7>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 8>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 4>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 1>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 2>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, 3>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl Instance<RegisterBlock, { crate::SOLE_INSTANCE }>

pub const unsafe fn instance() -> Self

Acquire a vaild, but possibly aliased, instance.

Safety

See the struct-level safety documentation.

impl<T, const N: u8> Instance<T, N>

pub const unsafe fn new(ptr: *const T) -> Self

Create an arbitrary Instance from a pointer to T.

Safety

See the struct docs for the safety contract.