Module imxrt_ral::tmr::CTRL0::CM::RW

Constants

CM_0

No operation

CM_1

Count rising edges of primary sourceRising edges are counted only when SCTRL[IPS] = 0. Falling edges are counted when SCTRL[IPS] = 1. If the primary count source is IP bus clock divide by 1, only rising edges are counted regardless of the value of SCTRL[IPS].

CM_2

Count rising and falling edges of primary sourceIP bus clock divide by 1 cannot be used as a primary count source in edge count mode.

CM_3

Count rising edges of primary source while secondary input high active

CM_4

Quadrature count mode, uses primary and secondary sources

CM_5

Count rising edges of primary source; secondary source specifies directionRising edges are counted only when SCTRL[IPS] = 0. Falling edges are counted when SCTRL[IPS] = 1.

CM_6

Edge of secondary source triggers primary count until compare

CM_7

Cascaded counter mode (up/down)The primary count source must be set to one of the counter outputs.