imxrt_ral::usb

Module ENDPTCTRL0

Endpoint Control0

Modules

RXE

RX Endpoint Enable 1 Enabled Endpoint0 is always enabled.

RXS

RX Endpoint Stall - Read/Write 0 End Point OK

RXT

RX Endpoint Type - Read/Write 00 Control Endpoint0 is fixed as a Control End Point.

TXE

TX Endpoint Enable 1 Enabled Endpoint0 is always enabled.

TXS

TX Endpoint Stall - Read/Write 0 End Point OK [Default] 1 End Point Stalled Software can write a one to this bit to force the endpoint to return a STALL handshake to the Host

TXT

TX Endpoint Type - Read/Write 00 - Control Endpoint0 is fixed as a Control End Point.