imxrt_ral::usb

Module HCCPARAMS

Host Controller Capability Parameters

Modules

ADC

64-bit Addressing Capability This bit is set ‘0b’ in all controller core, no 64-bit addressing capability is supported

ASP

Asynchronous Schedule Park Capability If this bit is set to a one, then the host controller supports the park feature for high-speed queue heads in the Asynchronous Schedule

EECP

EHCI Extended Capabilities Pointer

IST

Isochronous Scheduling Threshold

PFL

Programmable Frame List Flag If this bit is set to zero, then the system software must use a frame list length of 1024 elements with this host controller