imxrt_ral::usbnc
Module USB_OTG1_CTRL

USB OTG1 Control Register

Modules

OVER_CUR_DIS

Disable OTG1 Overcurrent Detection

OVER_CUR_POL

OTG1 Polarity of Overcurrent The polarity of OTG1 port overcurrent event

PWR_POL

OTG1 Power Polarity This bit should be set according to PMIC Power Pin polarity.

WIE

OTG1 Wake-up Interrupt Enable This bit enables or disables the OTG1 wake-up interrupt

WIR

OTG1 Wake-up Interrupt Request This bit indicates that a wake-up interrupt request is received on the OTG1 port

WKUP_DPDM_EN

Wake-up on DPDM change enable

WKUP_ID_EN

OTG1 Wake-up on ID change enable

WKUP_SW

OTG1 Software Wake-up

WKUP_SW_EN

OTG1 Software Wake-up Enable

WKUP_VBUS_EN

OTG1 wake-up on VBUS change enable