Module imxrt_ral::usbphy::STATUS

USB PHY Status Register

Modules

DEVPLUGIN_STATUS

Indicates that the device has been connected on the USB_DP and USB_DM lines.

HOSTDISCONDETECT_STATUS

Indicates that the device has disconnected while in high-speed host mode.

OTGID_STATUS

Indicates the results of ID pin on MiniAB plug

RESUME_STATUS

Indicates that the host is sending a wake-up after suspend and has triggered an interrupt.

RSVD0

Reserved.

RSVD1

Reserved.

RSVD2

Reserved.

RSVD3

Reserved.

RSVD4

Reserved.