Crate imxrt_ral

This project provides a register access layer (RAL) for all NXP i.mx rt microcontrollers.

When built, you must specify a device feature, such as imxrt1062. This will cause all modules in that device’s module to be re-exported from the top level, so that for example imxrt_ral::gpio will resolve to imxrt_ral::imxrt1062::gpio.

In the generated documentation, all devices are visible inside their family modules, but when built for a specific device, only that devices’ constants will be available.

Modules

imxrt101

Parent module for all IMXRT101 devices.

imxrt102

Parent module for all IMXRT102 devices.

imxrt105

Parent module for all IMXRT105 devices.

imxrt106

Parent module for all IMXRT106 devices.

Macros

modify_reg

Modify a RWRegister or UnsafeRWRegister.

read_reg

Read the value from a RORegister, RWRegister, UnsafeRORegister, or UnsafeRWRegister.

reset_reg

Reset a RWRegister, UnsafeRWRegister, WORegister, or UnsafeWORegister to its reset value.

write_reg

Write to a RWRegister or UnsafeRWRegister.

Structs

RORegister

A read-only register of type T.

RWRegister

A read-write register of type T.

UnsafeRORegister

A read-only register of type T, where read access is unsafe.

UnsafeRWRegister

A read-write register of type T, where read/write access is unsafe.

UnsafeWORegister

A write-only register of type T, where write access is unsafe.

WORegister

A write-only register of type T.

Enums

Interrupt

Interrupt sources