Module imxrt_ral::dcdc::REG2

DCDC Register 2

Modules

DCM_SET_CTRL

Set high to improve the transition from heavy load to light load

DISABLE_PULSE_SKIP

Set to “0” : stop charging if the duty cycle is lower than what set by dcdc_neglimit_in

LOOPCTRL_DC_C

Ratio of integral control parameter to proportional control parameter in the switching DC-DC converter, and can be used to optimize efficiency and loop response

LOOPCTRL_DC_FF

Two’s complement feed forward step in duty cycle in the switching DC-DC converter

LOOPCTRL_DC_R

Magnitude of proportional control parameter in the switching DC-DC converter control loop.

LOOPCTRL_EN_RCSCALE

Enable analog circuit of DC-DC converter to respond faster under transient load conditions.

LOOPCTRL_HYST_SIGN

Invert the sign of the hysteresis in DC-DC analog comparators.

LOOPCTRL_RCSCALE_THRSH

Increase the threshold detection for RC scale circuit.