imxrt_ral::dma
Module tcd

Modules

TCD_ATTR

TCD Transfer Attributes

TCD_BITER_ELINKNO

TCD Beginning Minor Loop Link, Major Loop Count (Channel Linking Disabled)

TCD_CITER_ELINKNO

TCD Current Minor Loop Link, Major Loop Count (Channel Linking Disabled)

TCD_CSR

TCD Control and Status

TCD_DADDR

TCD Destination Address

TCD_DLASTSGA

TCD Last Destination Address Adjustment/Scatter Gather Address

TCD_DOFF

TCD Signed Destination Address Offset

TCD_NBYTES_MLNO

TCD Minor Byte Count (Minor Loop Mapping Disabled)

TCD_SADDR

TCD Source Address

TCD_SLAST

TCD Last Source Address Adjustment

TCD_SOFF

TCD Signed Source Address Offset

Structs

RegisterBlock

Cluster TCD%s, containing TCD*_SADDR, TCD*_SOFF, TCD*_ATTR, TCD*_NBYTES_MLNO, TCD*_NBYTES_MLOFFNO, TCD*_NBYTES_MLOFFYES, TCD*_SLAST, TCD*_DADDR, TCD*_DOFF, TCD*_CITER_ELINKNO, TCD*_CITER_ELINKYES, TCD*_DLASTSGA, TCD*_CSR, TCD*_BITER_ELINKNO, TCD*_BITER_ELINKYES