Module imxrt_ral::flexspi::INTEN

Interrupt Enable Register

Modules

AHBBUSTIMEOUTEN

AHB Bus timeout interrupt.Refer Interrupts chapter for more details.

AHBCMDERREN

AHB triggered Command Sequences Error Detected interrupt enable.

AHBCMDGEEN

AHB triggered Command Sequences Grant Timeout interrupt enable.

IPCMDDONEEN

IP triggered Command Sequences Execution finished interrupt enable.

IPCMDERREN

IP triggered Command Sequences Error Detected interrupt enable.

IPCMDGEEN

IP triggered Command Sequences Grant Timeout interrupt enable.

IPRXWAEN

IP RX FIFO WaterMark available interrupt enable.

IPTXWEEN

IP TX FIFO WaterMark empty interrupt enable.

SCKSTOPBYRDEN

SCK is stopped during command sequence because Async RX FIFO full interrupt enable.

SCKSTOPBYWREN

SCK is stopped during command sequence because Async TX FIFO empty interrupt enable.

SEQTIMEOUTEN

Sequence execution timeout interrupt enable.Refer Interrupts chapter for more details.