Module imxrt_ral::flexspi::INTR

Interrupt Register

Modules

AHBBUSTIMEOUT

AHB Bus timeout interrupt.Refer Interrupts chapter for more details.

AHBCMDERR

AHB triggered Command Sequences Error Detected interrupt. When an error detected for AHB command, this command will be ignored and not executed at all.

AHBCMDGE

AHB triggered Command Sequences Grant Timeout interrupt.

IPCMDDONE

IP triggered Command Sequences Execution finished interrupt. This interrupt is also generated when there is IPCMDGE or IPCMDERR interrupt generated.

IPCMDERR

IP triggered Command Sequences Error Detected interrupt. When an error detected for IP command, this command will be ignored and not executed at all.

IPCMDGE

IP triggered Command Sequences Grant Timeout interrupt.

IPRXWA

IP RX FIFO watermark available interrupt.

IPTXWE

IP TX FIFO watermark empty interrupt.

SCKSTOPBYRD

SCK is stopped during command sequence because Async RX FIFO full interrupt.

SCKSTOPBYWR

SCK is stopped during command sequence because Async TX FIFO empty interrupt.

SEQTIMEOUT

Sequence execution timeout interrupt.