imxrt_ral::flexspi

Module INTR

Interrupt Register

Modules

AHBBUSERROR

AHB Bus timeout or AHB bus illegal access Flash during OTFAD key blob processing interrupt.

AHBCMDERR

AHB triggered Command Sequences Error Detected interrupt. When an error detected for AHB command, this command will be ignored and not executed at all.

AHBCMDGE

AHB triggered Command Sequences Grant Timeout interrupt.

IPCMDDONE

IP triggered Command Sequences Execution finished interrupt. This interrupt is also generated when there is IPCMDGE or IPCMDERR interrupt generated.

IPCMDERR

IP triggered Command Sequences Error Detected interrupt. When an error detected for IP command, this command will be ignored and not executed at all.

IPCMDGE

IP triggered Command Sequences Grant Timeout interrupt.

IPRXWA

IP RX FIFO watermark available interrupt.

IPTXWE

IP TX FIFO watermark empty interrupt.

KEYDONE

OTFAD key blob processing done interrupt.

KEYERROR

OTFAD key blob processing error interrupt.

SCKSTOPBYRD

SCLK is stopped during command sequence because Async RX FIFO full interrupt.

SCKSTOPBYWR

SCLK is stopped during command sequence because Async TX FIFO empty interrupt.

SEQTIMEOUT

Sequence execution timeout interrupt.