Module imxrt_ral::flexspi::MCR0

Module Control Register 0

Modules

AHBGRANTWAIT

Timeout wait cycle for AHB command grant.

ARDFEN

Enable AHB bus Read Access to IP RX FIFO.

ATDFEN

Enable AHB bus Write Access to IP TX FIFO.

COMBINATIONEN

This bit is to support Flash Octal mode access by combining Port A and B Data pins (SIOA[3:0] and SIOB[3:0]).

DOZEEN

Doze mode enable bit

HSEN

Half Speed Serial Flash access Enable.

IPGRANTWAIT

Time out wait cycle for IP command grant.

MDIS

Module Disable

RXCLKSRC

Sample Clock source selection for Flash Reading

SCKFREERUNEN

This bit is used to force SCK output free-running. For FPGA applications, external device may use SCK clock as reference clock to its internal PLL. If SCK free-running is enabled, data sampling with loopback clock from SCK pad is not supported (MCR0[RXCLKSRC]=2).

SWRESET

Software Reset