imxrt_ral::flexspi

Module MCR2

Module Control Register 2

Modules

CLRAHBBUFOPT

This bit determines whether AHB RX Buffer and AHB TX Buffer will be cleaned automatically when FlexSPI returns STOP mode ACK. Software should set this bit if AHB RX Buffer or AHB TX Buffer will be powered off in STOP mode. Otherwise AHB read access after exiting STOP mode may hit AHB RX Buffer or AHB TX Buffer but their data entries are invalid.

CLRLEARNPHASE

The sampling clock phase selection will be reset to phase 0 when this bit is written with 0x1. This bit will be auto-cleared immediately.

RESUMEWAIT

Wait cycle (in AHB clock cycle) for idle state before suspended command sequence resumed.

SAMEDEVICEEN

All external devices are same devices (both in types and size) for A1/A2/B1/B2.

SCKBDIFFOPT

B_SCLK pad can be used as A_SCLK differential clock output (inverted clock to A_SCLK). In this case, port B flash access is not available. After changing the value of this field, MCR0[SWRESET] should be set.