Module imxrt_ral::imxrt105::peripherals::snvs::LPLR

SNVS_LP Lock Register

Modules

GPR_HL

General Purpose Register Hard Lock When set, prevents any writes to the GPR

LPCALB_HL

LP Calibration Hard Lock When set, prevents any writes to the LP Calibration Value (LPCALB_VAL) and LP Calibration Enable (LPCALB_EN)

LPSVCR_HL

LP Security Violation Control Register Hard Lock When set, prevents any writes to the LPSVCR

LPTDCR_HL

LP Tamper Detectors Configuration Register Hard Lock When set, prevents any writes to the LPTDCR

MC_HL

Monotonic Counter Hard Lock When set, prevents any writes (increments) to the MC Registers and MC_ENV bit

MKS_HL

Master Key Select Hard Lock When set, prevents any writes to the MASTER_KEY_SEL field of the LP Master Key Control Register

SRTC_HL

Secure Real Time Counter Hard Lock When set, prevents any writes to the SRTC registers, SRTC_ENV, and SRTC_INV_EN bits

ZMK_RHL

Zeroizable Master Key Read Hard Lock When set, prevents any software reads to the ZMK registers and ZMK_ECC_VALUE field of the LPMKCR

ZMK_WHL

Zeroizable Master Key Write Hard Lock When set, prevents any writes (software and hardware) to the ZMK registers and ZMK_HWP, ZMK_VAL, and ZMK_ECC_EN fields of the LPMKCR