Module imxrt_ral::imxrt106::peripherals::pxp::AS_CTRL

Alpha Surface Control

Modules

ALPHA

Alpha modifier used when the ALPHA_MULTIPLY or ALPHA_OVERRIDE values are programmed in PXP_AS_CTRL[ALPHA_CTRL]

ALPHA_CTRL

Determines how the alpha value is constructed for this alpha surface

ALPHA_INVERT

Setting this bit to logic 0 will not alter the alpha value

ENABLE_COLORKEY

Indicates that colorkey functionality is enabled for this alpha surface

FORMAT

Indicates the input buffer format for AS.

ROP

Indicates a raster operation to perform when enabled

RSVD0

Reserved, always set to zero.

RSVD1

Reserved, always set to zero.