imxrt_ral::lcdif

Module CTRL2_SET

LCDIF General Control2 Register

Modules

BURST_LEN_8

By default, when the LCDIF is in the bus master mode, it will issue AXI bursts of length 16 (except when in packed 24 bpp mode, it will issue bursts of length 15)

EVEN_LINE_PATTERN

This field determines the order of the RGB components of each pixel in EVEN lines (line numbers 2,4,6,

ODD_LINE_PATTERN

This field determines the order of the RGB components of each pixel in ODD lines (line numbers 1,3,5,

OUTSTANDING_REQS

This bitfield indicates the maximum number of outstanding transactions that LCDIF should request when it is acting as a bus master