Macro imxrt_ral::read_reg

macro_rules! read_reg {
    ($periph : path, $instance : expr, $reg : ident, $($field : ident), +) => { ... };
    ($periph : path, $instance : expr, $reg : ident, $field : ident $($cmp : tt)
 *) => { ... };
    ($periph : path, $instance : expr, $reg : ident) => { ... };
}

Read the value from a RORegister, RWRegister, UnsafeRORegister, or UnsafeRWRegister.

Examples

// Safely acquire the peripheral instance (will panic if already acquired)
let gpio = imxrt_ral::gpio::GPIO1::take().unwrap();

// Read the whole register.
let val = read_reg!(imxrt_ral::gpio, gpio, ICR2);

// Read one field from the register.
let val = read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR29);

// Read multiple fields from the register.
let (val1, val2, val3) = read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR16, ICR21, ICR29);

// Check if one field is equal to a specific value, with the field's named values in scope.
while read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR29 == HIGH_LEVEL) {}

// Unsafe access without requiring you to first `take()` the instance
let val = unsafe { read_reg!(imxrt_ral::gpio, GPIO1, ICR2) };

Usage

Like write_reg!, this macro can be used multiple ways, either reading the entire register or reading a one or more fields from it and potentially performing a comparison with one field.

In all cases, the first arguments are:

the path to the peripheral module: imxrt_ral::gpio,

a reference to the instance of that peripheral: ‘gpio’ (anything which dereferences to RegisterBlock, such as Instance, &Instance, &RegisterBlock, or *const RegisterBlock),

the register you wish to access: ICR2 (a field on the RegisterBlock).

In the whole-register usage, the macro simply returns the register’s value:

// Read the entire value of GPIO1.ICR2 into `val`.
let val = read_reg!(imxrt_ral::gpio, gpio, ICR2);

For reading individual fields, the macro masks and shifts appropriately:

// Read just the value of the field GPIO1.ICR29 into `val`.
let val = read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR29);

// As above, but expanded for exposition:
let val = ((*gpio).ICR2.read() & imxrt_ral::gpio::ICR2::ICR29::mask)
          >> imxrt_ral::gpio::ICR2::ICR29::offset;

// Read multiple fields
let (val1, val2) = read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR29, ICR30);

// As above, but expanded for exposition:
let (val1, val2) = { let val = (*gpio).ICR2.read();
    ((val & imxrt_ral::gpio::ICR2::ICR29::mask) >> imxrt_ral::gpio::ICR2::ICR29::offset,
     (val & imxrt_ral::gpio::ICR2::ICR30::mask) >> imxrt_ral::gpio::ICR2::ICR30::offset,
    )};

For comparing a single field, the macro masks and shifts and then performs the comparison:

// Loop while ICR29 is HIGH_LEVEL.
while read_reg!(imxrt_ral::gpio, gpio, ICR2, ICR29 == HIGH_LEVEL) {}

Safety

This macro will require an unsafe function or block when used with an UnsafeRWRegister or UnsafeRORegister, but not if used with RWRegister, or RORegister.

When run in an unsafe context, peripheral instances are directly accessible without requiring having called take() beforehand:

let val = unsafe { read_reg!(imxrt_ral::gpio, GPIO1, ICR1) };

This works because GPIO1 is a *const RegisterBlock in the imxrt_ral::gpio module; and the macro brings such constants into scope and then dereferences the provided reference.