imxrt_ral::ocotp

Module LOCK

Value of OTP Bank0 Word0 (Lock controls)

Modules

ANALOG

Status of shadow register and OTP write lock for analog region

BOOT_CFG

Status of shadow register and OTP write lock for boot_cfg region

FIELD_RETURN

Reserved

GP1

Status of shadow register and OTP write lock for gp1 region

GP2

Status of shadow register and OTP write lock for gp2 region

GP3

Status of shadow register and OTP write lock for gp3 region

MAC_ADDR

Status of shadow register and OTP write lock for mac_addr region

MEM_TRIM

Status of shadow register and OTP write lock for mem_trim region

MISC_CONF

Status of shadow register and OTP write lock for misc_conf region

OTPMK_CRC

Status of shadow register and OTP write lock for otpmk_crc region

OTPMK_LSB

Status of shadow register read and write, OTP read and write lock for otpmk region (LSB)

OTPMK_MSB

Status of shadow register read and write, OTP read and write lock for otpmk region (MSB)

SJC_RESP

Status of shadow register read and write, OTP read and write lock for sjc_resp region

SW_GP1

Status of shadow register and OTP write lock for sw_gp1 region

SW_GP2_LOCK

Status of shadow register and OTP write lock for sw_gp2 region

SW_GP2_RLOCK

Status of shadow register and OTP read lock for sw_gp2 region

TESTER

Status of shadow register and OTP write lock for tester region