Module imxrt_ral::pmu::REG_2P5_SET

Regulator 2P5 Register

Modules

BO_OFFSET

Control bits to adjust the regulator brownout offset voltage in 25mV steps

BO_VDD2P5

Status bit that signals when a brownout is detected on the regulator output.

ENABLE_BO

Control bit to enable the brownout circuitry in the regulator.

ENABLE_ILIMIT

Control bit to enable the current-limit circuitry in the regulator.

ENABLE_LINREG

Control bit to enable the regulator output.

ENABLE_PULLDOWN

Control bit to enable the pull-down circuitry in the regulator

ENABLE_WEAK_LINREG

Enables the weak 2p5 regulator

OK_VDD2P5

Status bit that signals when the regulator output is ok. 1 = regulator output > brownout target

OUTPUT_TRG

Control bits to adjust the regulator output voltage