imxrt_ral::snvs::HPCR::PI_FREQ

Module RW

Constants

PI_FREQ_0

bit 0 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_1

bit 1 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_2

bit 2 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_3

bit 3 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_4

bit 4 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_5

bit 5 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_6

bit 6 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_7

bit 7 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_8

bit 8 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_9

bit 9 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_10

bit 10 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_11

bit 11 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_12

bit 12 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_13

bit 13 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_14

bit 14 of the HPRTCLR is selected as a source of the periodic interrupt

PI_FREQ_15

bit 15 of the HPRTCLR is selected as a source of the periodic interrupt