imxrt_ral::snvs

Module HPLR

SNVS_HP Lock Register

Modules

GPR_SL

General Purpose Register Soft Lock When set, prevents any writes to the GPR

HAC_L

High Assurance Counter Lock When set, prevents any writes to HPHACIVR, HPHACR, and HAC_EN bit of HPCOMR

HPSICR_L

HP Security Interrupt Control Register Lock When set, prevents any writes to the HPSICR

HPSVCR_L

HP Security Violation Control Register Lock When set, prevents any writes to the HPSVCR

LPCALB_SL

LP Calibration Soft Lock When set, prevents any writes to the LP Calibration Value (LPCALB_VAL) and LP Calibration Enable (LPCALB_EN)

LPSVCR_SL

LP Security Violation Control Register Soft Lock When set, prevents any writes to the LPSVCR

LPTDCR_SL

LP Tamper Detectors Configuration Register Soft Lock When set, prevents any writes to the LPTDCR

MC_SL

Monotonic Counter Soft Lock When set, prevents any writes (increments) to the MC Registers and MC_ENV bit

MKS_SL

Master Key Select Soft Lock When set, prevents any writes to the MASTER_KEY_SEL field of the LPMKCR

SRTC_SL

Secure Real Time Counter Soft Lock When set, prevents any writes to the SRTC Registers, SRTC_ENV, and SRTC_INV_EN bits

ZMK_RSL

Zeroizable Master Key Read Soft Lock When set, prevents any software reads to the ZMK Registers and ZMK_ECC_VALUE field of the LPMKCR

ZMK_WSL

Zeroizable Master Key Write Soft Lock When set, prevents any writes (software and hardware) to the ZMK registers and the ZMK_HWP, ZMK_VAL, and ZMK_ECC_EN fields of the LPMKCR