Module imxrt_ral::snvs::HPSR

SNVS_HP Status Register

Modules

BI

Button Interrupt Signal ipi_snvs_btn_int_b was asserted.

BTN

Button Value of the BTN input

HPTA

HP Time Alarm Indicates that the HP Time Alarm has occurred since this bit was last cleared.

LPDIS

Low Power Disable If 1, the low power section has been disabled by means of an input signal to SNVS

OTPMK_SYNDROME

One Time Programmable Master Key Syndrome In the case of a single-bit error, the eight lower bits of this value indicate the bit number of error location

OTPMK_ZERO

One Time Programmable Master Key is Equal to Zero

PI

Periodic Interrupt Indicates that periodic interrupt has occurred since this bit was last cleared.

SECURITY_CONFIG

Security Configuration This field reflects the settings of the sys_secure_boot input and the three security configuration inputs to SNVS

SSM_STATE

System Security Monitor State This field contains the encoded state of the SSM’s state machine

ZMK_ZERO

Zeroizable Master Key is Equal to Zero