imxrt_ral::snvs

Module HPSVCR

SNVS_HP Security Violation Control Register

Modules

LPSV_CFG

LP Security Violation Configuration This field configures the LP security violation source.

SV0_CFG

Security Violation 0 Security Violation Configuration This field configures the Security Violation 0 Security Violation Input

SV1_CFG

Security Violation 1 Security Violation Configuration This field configures the Security Violation 1 Security Violation Input

SV2_CFG

Security Violation 2 Security Violation Configuration This field configures the Security Violation 2 Security Violation Input

SV3_CFG

Security Violation 3 Security Violation Configuration This field configures the Security Violation 3 Security Violation Input

SV4_CFG

Security Violation 4 Security Violation Configuration This field configures the Security Violation 4 Security Violation Input

SV5_CFG

Security Violation 5 Security Violation Configuration This field configures the Security Violation 5 Security Violation Input