imxrt_ral::snvs
Module LPMKCR

SNVS_LP Master Key Control Register

Modules

MASTER_KEY_SEL

Master Key Select These bits select the SNVS Master Key output when Master Key Select bits are enabled by MKS_EN bit in the HPCOMR

ZMK_ECC_EN

Zeroizable Master Key Error Correcting Code Check Enable Writing one to this field automatically calculates and sets the ZMK ECC value in the ZMK_ECC_VALUE field of this register

ZMK_ECC_VALUE

Zeroizable Master Key Error Correcting Code Value This field is automatically calculated and set when one is written into ZMK_ECC_EN bit of this register

ZMK_HWP

Zeroizable Master Key hardware Programming mode When set, only the hardware key programming mechanism can set the ZMK and software cannot read it

ZMK_VAL

Zeroizable Master Key Valid When set, the ZMK value can be selected by the master key control block for use by cryptographic modules