imxrt_ral

Trait Valid

pub trait Valid: Sealed { }

Vouches for an Instance<T, N>’s validity.

This trait is implemented for all Instance<T, N> supported by your chip. Note that the implementation may change when selecting new chip features. For instance, i.MX RT 1011 chips do not have LPUART 4 through 8. So, Valid is not implemented for lpuart::Instance<4> through lpuart::Instance<8>.

See the package documentation for more information on how to use this trait in your APIs.